### 2023-03-25 #93 @sampottinger
afscgap: Ocean health tools for NOAA AFSC GAP species presence data
**Submitting Author:** Sam Pottinger (@sampottinger)
**All current maintainers:** @sampottinger, @gizarp
**Package Name:** afscgap
**One-Line Description of Package:** Community contributed Python-based tools for working with public bottom trawl surveys data from the NOAA Alaska Fisheries Science Center Groundfish Assessment Program ([NOAA AFSC GAP](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program)).
**Repository Link:** https://github.com/SchmidtDSE/afscgap
**Version submitted:** 0.0.7
**Editor:** [Filipe Fernandes](https://github.com/ocefpaf)
**Reviewer 1:** [Tylar Murray](https://github.com/7yl4r)
**Reviewer 2:** [Ayush Anand](https://github.com/ayushanand18)
**Archive:** TBD
**Version accepted:** TBD
**Date accepted (month/day/year):** TBD

---

Python-based tool set for interacting with [bottom trawl surveys](https://www.fisheries.noaa.gov/alaska/commercial-fishing/alaska-groundfish-bottom-trawl-survey-data) from the [Ground Fish Assessment Program (GAP)](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program). This provides information about where certain species were seen and when under what conditions, information useful for research in ocean health.

It offers:

 - Pythonic access to the official [NOAA AFSC GAP API service](https://www.fisheries.noaa.gov/foss/f?p=215%3A28).
 - Memory-efficient tools for inference of the ["negative" observations](https://github.com/SchmidtDSE/afscgap#absence-vs-presence-data) not provided by the API service but required for some common types of analysis.
 - [Visualization tools](https://app.pyafscgap.org) for quickly exploring and creating comparisons within the dataset that provide an "on-ramp" to deeper analysis. The visual analytics components aim to serve both expert programmers and audiences with limited programming experience.

Note that GAP is an excellent dataset produced by the [Resource Assessment and Conservation Engineering (RACE) Division](https://www.fisheries.noaa.gov/about/resource-assessment-and-conservation-engineering-division) of the [Alaska Fisheries Science Center (AFSC)](https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center) as part of the National Oceanic and Atmospheric Administration's Fisheries organization ([NOAA Fisheries](https://www.fisheries.noaa.gov/)). See also the [RACEBASE NOAA InPort entry](https://www.fisheries.noaa.gov/inport/item/22008.).

Additional information at [https://pyafscgap.org/](https://pyafscgap.org/).

Check out our [package scope page][PackageCategories] to learn more about our scope. (If you are unsure of which category you fit, we suggest you make a pre-submission inquiry):

    - [X] Data visualization - [see presubmission](https://github.com/pyOpenSci/software-submission/issues/91)

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

This library supports retrieval of data from the official [NOAA AFSC GAP REST API service](https://apps-st.fisheries.noaa.gov/ods/foss/afsc_groundfish_survey/) but, as that service on its own is often not sufficient due to certain data limitations, it also offers [zero catch inference](https://pyafscgap.org/devdocs/afscgap.html#absence-vs-presence-data) as required for a number of common types of investigations (hence data processing). Finally, given the needs of the community and the vast breadth of the dataset, it offers a [community application](https://app.pyafscgap.org/) to explore the data which, in turn, can generate Python code to help users get started with continued analysis within their own scripts.

This project largely benefits scientific researchers in the ocean health space as this dataset is useful for fisheries management, biodiversity research, and marine science more generally. An example of what this analysis may look like is provided in our [example notebook](https://mybinder.org/v2/gh/SchmidtDSE/afscgap/main?urlpath=/tree/index.ipynb) hosted on mybinder.org.

We are not aware of other Python packages working with the AFSC GAP dataset.

Please see our [presubmission](https://github.com/pyOpenSci/software-submission/issues/91).

See [NOAA InPort entry](https://www.fisheries.noaa.gov/inport/item/22008). We are primarily using Distribution 6 in the API form. As described, these data are available with the following access constraints: "There are no legal restrictions on access to the data. They reside in public domain and can be freely distributed."

See [BSD License](https://github.com/SchmidtDSE/afscgap/blob/main/LICENSE.md) within repo.

See [project README.md](https://github.com/SchmidtDSE/afscgap/blob/main/LICENSE.md).

See [project website](https://pyafscgap.org/) and [usage section of project website's documentation microsite](https://pyafscgap.org/devdocs/afscgap.html#usage).

See [tutorial notebook on mybinder](https://mybinder.org/v2/gh/SchmidtDSE/afscgap/main?urlpath=/tree/index.ipynb).

See [main tests for package](https://github.com/SchmidtDSE/afscgap/tree/main/afscgap/test) and supplemental tests for website ([backend tests](https://github.com/SchmidtDSE/afscgap/tree/main/afscgapviz/test) and [frontend tests](https://github.com/SchmidtDSE/afscgap/tree/main/afscgapviz/static/test)).

See [CI](https://github.com/SchmidtDSE/afscgap/actions/workflows/build.yml) and CD ([library](https://github.com/SchmidtDSE/afscgap/actions/workflows/deploy.yml), [documentation](https://github.com/SchmidtDSE/afscgap/actions/workflows/docs.yml)).

See [example notebook](https://mybinder.org/v2/gh/SchmidtDSE/afscgap/main?urlpath=/tree/index.ipynb) and [app intro](https://app.pyafscgap.org/).

In addition to query compilation / emulation, we are hopeful that the negative record inference and complementing visualization tool are enough to escape the thin API client status. Coming in at over 10k lines of code, we would seem to fit some of the ["hard" criteria that the journal puts forward](https://joss.readthedocs.io/en/latest/submitting.html). Though this library happily provides Pythonic access to an API service, we believe that this contribution as a whole goes beyond a thin client by providing one of the only public mechanisms for conducting investigation requiring negative catch data and provides unique tools for comparative analysis within with the dataset.

See https://github.com/SchmidtDSE/afscgap/blob/main/inst/paper.md (or PDF preview at https://github.com/SchmidtDSE/afscgap/blob/main/inst/paper.pdf).

We have submitted to Code Ocean. Please confirm if this will suffice. See 10.24433/CO.4905407.v1 / https://codeocean.com/capsule/4905407/tree/v1

We would be delighted to have your pull requests! 🎉

- [X] Last but not least please fill out our pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us track

[Templates]: https://www.pyopensci.org/software-peer-review/appendices/templates.html
