+++ inst/paper.bib
@@ -108,7 +108,7 @@
 }
 
  @misc{hayashida,
-  title={The Structure of Fun},
+  title={The Structure of Fun: Learning from Super Mario 3D Land's Director},
   url={https://www.gamedeveloper.com/design/the-structure-of-fun-learning-from-i-super-mario-3d-land-i-s-director},
   journal={Game Developer},
   publisher={Informa},
