+++ inst/paper.bib
@@ -24,7 +24,7 @@
   year={2013}
 } 
 
-@misc{decorator,
+@misc{decorators,
   title={Decorator},
   url={https://refactoring.guru/design-patterns/decorator},
   journal={Refactoring.Guru},
