### 2023-03-27 #93 @NickleDave
Hi again @sampottinger I'm adding the initial editor checks below.

Good news! afscgap passes with flying colors. I made a couple of comments in the checklist but I don't see anything that needs to be addressed--looks like we're ready for review.

I will begin looking for an editor this week and update you here as soon as we have found one.

  - might be worth adding a conda-forge package?
  - [x] The package imports properly into a standard Python environment `import package-name`.
  - yes, data retrieval (from GAP dataset through API), data processing, data visualization
    - https://pyafscgap.org/devdocs/afscgap.html
    - notebooks etc linked from landing page https://pyafscgap.org (that also links to install / API docs)
  - [x] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
  - [x] **Code of Conduct** The package has a `Code of Conduct` file.
    - adapted from p5.js COC
    - BSD 3-Clause
- [x] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.

---
---
