### 2023-04-03 d582cd0
Fix missing note
Fix a missing note in viz section that might have been lost in editing.
