### 2023-04-03 6ddc1ef
Update paper.md for side by side
