### 2023-04-03 f04b6cb
Add few citation to bib file.
