+++ inst/paper.bib
@@ -69,6 +69,16 @@
   year={2023}
 }
 
+@misc{few,
+  title={Coordinated highlighting in context},
+  url={https://www.perceptualedge.com/articles/visual_business_intelligence/coordinated_highlighting_in_context.pdf},
+  journal={Visual Business Intelligence Newsletter},
+  publisher={Perceptual Edge},
+  author={Few, Stephen},
+  year={2010},
+  month={Sep}
+} 
+
 @misc{flask,
   title={Welcome to Flask},
   url={https://flask.palletsprojects.com/en/2.2.x/},
