### 2023-04-04 0e587dc
Add paper to CI / CD.
