### 2023-04-04 fc153ca
Fixes for #47.
