+++ inst/README.md
@@ -4,3 +4,5 @@ Resources for the JOSS paper submission for the Pyafscgap.org project. See paper
 
  - [pandoc/luafilters](https://github.com/pandoc/lua-filters) by pandoc Lua filters contributors under the [MIT License](https://github.com/pandoc/lua-filters/blob/master/LICENSE)
  - [pandoc](https://pandoc.org/) (via call to OS package through shell script) by John MacFarlane, Albert Krewinkel, Jesse Rosenthal, and other contributors under the [GPL License](https://github.com/jgm/pandoc#license).
+
+The paper preview can be built by running `bash preview_paper.sh`.
