### 2023-04-10 #93 @sampottinger
Thanks again @ocefpaf for your thoughtful comments! Please see below.

> I see you have a module named utils

This is a good suggestion. I made a PR at https://github.com/SchmidtDSE/afscgap/pull/49. Let me know if this refactor looks good to you!

> The package name suffers from the usual NOAA acronym problem ... if there is a user base, even a small one, please disregard this comment

Yeah the dataset's whole name is NOAA AFSC GAP RACEBACSE FOSS :sweat_smile: . Thanks for being understanding on this one. I think we will politely decline this suggestion if that's OK. We have some community continuity concerns like you mentioned and some community sensitivity concerns we want to respect... we have to be a little careful around not appearing to have activity in other datasets (NOAA or otherwise) that may be at different moments in their open science journey than AFSC GAP or not (yet? :crossed_fingers:) open in the same way that the AFSC GAP API service is. Given our current context, this specificity is probably desirable for at least the medium-term future.

> a nice addition to our Code Lab

Amazing! We'd be honored :smile:. We'd be more than happy to co-develop something with you or another community member in that space. Just for the conversation, I'll highlight that we do have a [mybinder](https://mybinder.org/v2/gh/SchmidtDSE/afscgap/main?urlpath=/tree/index.ipynb) example available ([source in repo](https://github.com/SchmidtDSE/afscgap/blob/main/notebooks/cod.ipynb)). Totally understand if it might not be quite the right vibe though.

> The query method argument list is daunting!

I think that design choice is a strong opinion loosely held. I'm not sure if it is helpful but I'll document some thoughts we had internally but open to additional ideas!

We went this route to fulfill a few goals: we wanted to reflect the structure of the API service (any combination of filters in the keyword set is valid in the NOAA API service so there isn't a clear functional decomposition) while also providing documented discoverability. There were some other ideas explored:

- **kwargs**: Leaving the keyword arguments as variable is a bit easier for implementation but, though documentation for the AFSC GAP dataset exists, it's not necessarily convenient to reference / use from a Python context. So, we wanted to have docstring description for all of the available options to benefit IDEs, API doc generation, etc. We also wanted the possible list of filters to be enumerated for the user in those contexts as well.
- **pass objects**: There's the option of a more ORM-like approach where a list of filters are passed like `query(filters=[CommonNameFilter(eq='Pacific cod')]` but we didn't like forcing a lot of our type system into client code as it would require quite a large number of types to maintain our goals around docstring coverage of the available filter parameters and discoverability.
- **builder chaining**: Probably my personal favorite, we could turn this into an object with syntax like `query().commonName(eq='Pacific cod').year(min=2021).get()`. While I personally like this syntax, there's been conversation internally and with community partners for using this library in an educational context as well. As this syntax may be less common in earlier Python programming journeys, there was a feeling that the keyword arguments approach was more inclusive / approachable for someone in, say, an introductory Python course or a student not in a computer science curriculum (we got invited for one coming up soon!).

Just wanted to list this out but open to other thoughts! This just felt like the right balance around the broader goals with regards to documentation / discoverability, keeping the type system encapsulated (see below), and maintaining accessibility.

![library](https://user-images.githubusercontent.com/110391/230933367-d928115e-ba40-4f3e-b390-6a6ed6059845.png)
