### 2023-04-13 #93 @ocefpaf
Thanks [Tylar Murray](https://github.com/) and [Ayush Anand](https://github.com/ayushanand18) for volunteering to review afscgap!

@sampottinger both reviewers are experienced Scientific Python programmers and work on a similar project to make OBIS data easily accessible from Python. Hopefully both sides can benefit from this experience.

We usually aim for 3 weeks for review, 2 weeks for subsequent changes, and 1 week for reviewer approval of changes. @7yl4r and ayushanand18 please let me know if you have any questions. You can comment, open issue, and/or PR directly in the afscgap for this review. Just make sure to cross-reference them here.

Thank you both again!
