### 2023-04-13 #93 @sampottinger
Thanks @ocefpaf!

> Better than I thought.

Awesome! Merged in.

> I disagree with the learning curve there.

That's helpful to have your perspective on that. Let me float this around a little bit and see how we feel about migration. I'll try to make a call on that in the next day or so to keep things moving.

> With that said, again, anything I write here is not a blocker for acceptance. The current implementation is good and solid. Just suggestions for the future.

Thanks!

> wave Hi [Tylar Murray](https://github.com/7yl4r) and [Ayush Anand](https://github.com/ayushanand18)!

Lovely to meet you both! Thank you for your time. :D
