### 2023-04-14 #93 @ayushanand18
Thank you so much @ocefpaf and @sampottinger!
Great work building this package. The current implementation looks robust and clean.

But I do have some observations to make.
1. There is this function `afscgap.query().get_bottom_temperature_c()`, and I do think that it might be hard to guess the function name for someone who hasn't seen this dataset before. Mind renaming to something like `get_bottom_temperature(unit="celsius"|'c')`? This observation is similar to other functions like `cpue_*()`, `duration_hr`, etc. I think could have been derived from the direct field names in the dataset, and those who know about the dataset can easily figure out what the name would be, without looking at the docs, but it would be difficult for people from outside (like me) to figure out without going back-and-forth to the docs.

2. Also about the filters and getters, we can utilise the __getitem__() method in the class with something like this,
```python
def __getitem__(self, key):
    return self.__getattribute__(key)
```
This might be more intuitive. Or to make it much easier can we can just go by the name `get(key)`, and specify the key we want. [Ref](https://github.com/SchmidtDSE/afscgap/blob/main/afscgap/client.py)

I would love to hear your opinion, and these are surely non-blockers for this review!
