### 2023-04-17 SchmidtDSE/afscgap#50 @sampottinger
Switch query to builder pattern
Part of https://github.com/pyOpenSci/software-submission/issues/93, thanks @ocefpaf. Example syntax desired:

```
query = afscgap.query()
query.filterCommonName(eq='Pacific cod')
query.filterYear(min=2021)
results = query.execute()
```

or

```
results = afscgap.query() \
  .filterCommonName(eq='Pacific cod') \
  .filterYear(min=2021) \
  .execute()
```

This achieves parity with other scientific python packages in terms of patterns.
