### 2023-04-17 SchmidtDSE/afscgap#51 @sampottinger
Consolate units behind getters
Part of https://github.com/pyOpenSci/software-submission/issues/93, thanks @ayushanand18. Example of desired behavior:

```
for result in results:
  result.get_bottom_temperature(units='c')
  result.get_bottom_temperature(units='f')
```

Should set defaults and document in the docstrings.
