### 2023-04-17 #93 @sampottinger
Hello @ayushanand18 and @ocefpaf!

> I do like the idea of the builder chaining and it is not too uncommon

Thanks again for the conversation @ocefpaf! I've started implementing the builder pattern at https://github.com/SchmidtDSE/afscgap/issues/50.

> Mind renaming to something like `get_bottom_temperature(unit="celsius"|'c')`

Thanks for the suggestion @ayushanand18! I've started that at https://github.com/SchmidtDSE/afscgap/issues/51

> we can utilise the getitem() method in the class

Thanks for this suggestion but I may pass on this if that's ok @ayushanand18.

Our reasoning is that the library carries an explicit goal to provide docstrings and type annotations for these fields. I think, given the complexity of the data model itself as well as the difficulty in using some of the upstream documentation from NOAA directly in the Python ecosystem, I want to leave those methods there for discoverability and encourage use of those typed interfaces unless the user "opts out" first.

To that end, there is a `to_dict` method on `afscgap.model.Record` for those that prefer that key-based approach or want to read into something like `pandas.DataFrame` but we hope requiring `to_dict` first helps inform the user that the object returned from a query does offer those other interfaces, information which may be missed if they simply try treating the result as a dictionary. In other words, I'd rather have the user first explicitly decline use of the default typed / strong interface given the benefits it offers for static checking and affordance.
