### 2023-04-17 3ce6525
First draft of solution to #50
