+++ afscgap/test/test_entry.py
@@ -29,55 +29,53 @@ class EntryPointNoInferenceTests(unittest.TestCase):
         )
 
     def test_query_primitive(self):
-        result = afscgap.query(
-            year=2021,
-            srvy='BSS',
-            requestor=self._mock_requestor
-        )
-        results = list(result)
+        query = afscgap.Query(requestor=self._mock_requestor)
+        query.filter_year(eq=2021)
+        query.filter_srvy(eq='BSS')
+        results = list(query.execute())
         self.assertEquals(len(results), 20)
 
     def test_query_dict(self):
-        result = afscgap.query(
-            year=2021,
-            latitude_dd={'$gte': 56.99, '$lte': 57.04},
-            longitude_dd={'$gte': -143.96, '$lte': -144.01},
-            requestor=self._mock_requestor
-        )
-        results = list(result)
+        query = afscgap.Query(requestor=self._mock_requestor)
+        query.filter_year(eq=2021)
+        query.filter_latitude_dd(eq={'$gte': 56.99, '$lte': 57.04})
+        query.filter_longitude_dd(eq={'$gte': -143.96, '$lte': -144.01})
+        results = list(query.execute())
+        self.assertEquals(len(results), 20)
+
+    def test_query_keywords(self):
+        query = afscgap.Query(requestor=self._mock_requestor)
+        query.filter_year(eq=2021)
+        query.filter_latitude_dd(min_val=56.99, max_val=57.04)
+        query.filter_longitude_dd(min_val=-143.96, max_val=-144.01)
+        results = list(query.execute())
         self.assertEquals(len(results), 20)
 
     def test_query_dict_filter_incomplete(self):
-        result = afscgap.query(
-            year=2021,
-            latitude_dd={'$gte': 56.99, '$lte': 57.04},
-            longitude_dd={'$gte': -143.96, '$lte': -144.01},
-            requestor=self._mock_requestor,
-            filter_incomplete=True
-        )
-        results = list(result)
+        query = afscgap.Query(requestor=self._mock_requestor)
+        query.filter_year(eq=2021)
+        query.filter_latitude_dd(eq={'$gte': 56.99, '$lte': 57.04})
+        query.filter_longitude_dd(eq={'$gte': -143.96, '$lte': -144.01})
+        query.set_filter_incomplete(True)
+        results = list(query.execute())
         self.assertEquals(len(results), 19)
 
     def test_query_dict_invalid_filter_incomplete(self):
-        result = afscgap.query(
-            year=2021,
-            latitude_dd={'$gte': 56.99, '$lte': 57.04},
-            longitude_dd={'$gte': -143.96, '$lte': -144.01},
-            requestor=self._mock_requestor,
-            filter_incomplete=True
-        )
-        list(result)
+        query = afscgap.Query(requestor=self._mock_requestor)
+        query.filter_year(eq=2021)
+        query.filter_latitude_dd(eq={'$gte': 56.99, '$lte': 57.04})
+        query.filter_longitude_dd(eq={'$gte': -143.96, '$lte': -144.01})
+        query.set_filter_incomplete(True)
+        result = query.execute()
         self.assertEquals(result.get_invalid().qsize(), 2)
 
     def test_query_dict_invalid_keep_incomplete(self):
-        result = afscgap.query(
-            year=2021,
-            latitude_dd={'$gte': 56.99, '$lte': 57.04},
-            longitude_dd={'$gte': -143.96, '$lte': -144.01},
-            requestor=self._mock_requestor,
-            filter_incomplete=False
-        )
-        list(result)
+        query = afscgap.Query(requestor=self._mock_requestor)
+        query.filter_year(eq=2021)
[…]
