+++ afscgap/__init__.py
@@ -32,6 +32,7 @@ from afscgap.typesdef import FLOAT_PARAM
 from afscgap.typesdef import INT_PARAM
 from afscgap.typesdef import STR_PARAM
 
+from afscgap.typesdef import OPT_FLOAT
 from afscgap.typesdef import OPT_INT
 from afscgap.typesdef import OPT_STR
 from afscgap.typesdef import OPT_REQUESTOR
@@ -46,244 +47,1027 @@ LARGE_WARNING = ' '.join([
 ])
 
 
-def query(
-    year: FLOAT_PARAM = None,
-    srvy: STR_PARAM = None,
-    survey: STR_PARAM = None,
-    survey_id: FLOAT_PARAM = None,
-    cruise: FLOAT_PARAM = None,
-    haul: FLOAT_PARAM = None,
-    stratum: FLOAT_PARAM = None,
-    station: STR_PARAM = None,
-    vessel_name: STR_PARAM = None,
-    vessel_id: FLOAT_PARAM = None,
-    date_time: STR_PARAM = None,
-    latitude_dd: FLOAT_PARAM = None,
-    longitude_dd: FLOAT_PARAM = None,
-    species_code: FLOAT_PARAM = None,
-    common_name: STR_PARAM = None,
-    scientific_name: STR_PARAM = None,
-    taxon_confidence: STR_PARAM = None,
-    cpue_kgha: FLOAT_PARAM = None,
-    cpue_kgkm2: FLOAT_PARAM = None,
-    cpue_kg1000km2: FLOAT_PARAM = None,
-    cpue_noha: FLOAT_PARAM = None,
-    cpue_nokm2: FLOAT_PARAM = None,
-    cpue_no1000km2: FLOAT_PARAM = None,
-    weight_kg: FLOAT_PARAM = None,
-    count: FLOAT_PARAM = None,
-    bottom_temperature_c: FLOAT_PARAM = None,
-    surface_temperature_c: FLOAT_PARAM = None,
-    depth_m: FLOAT_PARAM = None,
-    distance_fished_km: FLOAT_PARAM = None,
-    net_width_m: FLOAT_PARAM = None,
-    net_height_m: FLOAT_PARAM = None,
-    area_swept_ha: FLOAT_PARAM = None,
-    duration_hr: FLOAT_PARAM = None,
-    tsn: INT_PARAM = None,
-    ak_survey_id: INT_PARAM = None,
-    limit: OPT_INT = None,
-    start_offset: OPT_INT = None,
-    base_url: OPT_STR = None,
-    requestor: OPT_REQUESTOR = None,
-    filter_incomplete: bool = False,
-    presence_only: bool = True,
-    suppress_large_warning: bool = False,
-    hauls_url: OPT_STR = None,
-    warn_function: WARN_FUNCTION = None,
-    hauls_prefetch: OPT_HAUL_LIST = None) -> afscgap.cursor.Cursor:
-    """Execute a query against the AFSC GAP API.
+class Query:
+    """Entrypoint for the AFSC GAP Python library.
 
-    Args:
-        year: Filter on year for the survey in which this observation was made.
-            Pass None if no filter should be applied. Defaults to None.
-        srvy: Filter on the short name of the survey in which this observation
-            was made. Pass None if no filter should be applied. Defaults to
-            None. Note that common values include: NBS (N Bearing Sea), EBS (SE
-            Bearing Sea), BSS (Bearing Sea Slope), GOA (Gulf of Alaska), and
-            AI (Aleutian Islands).
-        survey: Filter on long form description of the survey in which the
-            observation was made. Pass None if no filter should be applied.
-            Defaults to None.
-        survey_id: Filter on unique numeric ID for the survey. Pass None if no
[…]
