### 2023-04-17 760979f
All but one test for #52.
