+++ afscgap/__init__.py
@@ -117,11 +117,11 @@ class Query:
         self._suppress_large_warning: bool = False
         self._warn_function: WARN_FUNCTION = None
         self._hauls_prefetch: OPT_HAUL_LIST = None
-    
+
     def filter_year(self, eq: FLOAT_PARAM = None, min_val: OPT_FLOAT = None,
         max_val: OPT_FLOAT = None) -> 'Query':
         """Filter on year for the survey in which this observation was made.
-        
+
         Args:
             eq: The exact value that must be matched for a record to be
                 returned. Pass None if no equality filter should be applied.
@@ -134,9 +134,9 @@ class Query:
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
         """
-        self._year = self._create_float_param(min_val, max_val, eq)
+        self._year = self._create_float_param(eq, min_val, max_val)
         return self
-    
+
     def filter_srvy(self, eq: STR_PARAM = None, min_val: OPT_STR = None,
         max_val: OPT_STR = None) -> 'Query':
         """Filter on haul survey short name.
@@ -146,7 +146,7 @@ class Query:
         that common values include: NBS (N Bearing Sea), EBS (SE Bearing Sea),
         BSS (Bearing Sea Slope), GOA (Gulf of Alaska), and AI (Aleutian
         Islands).
-        
+
         Args:
             eq: The exact value that must be matched for a record to be
                 returned. Pass None if no equality filter should be applied.
@@ -159,9 +159,9 @@ class Query:
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
         """
-        self._srvy = self._create_str_param(min_val, max_val, eq)
+        self._srvy = self._create_str_param(eq, min_val, max_val)
         return self
-    
+
     def filter_survey(self, eq: STR_PARAM = None, min_val: OPT_STR = None,
         max_val: OPT_STR = None) -> 'Query':
         """Filter on survey long name.
@@ -181,9 +181,9 @@ class Query:
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
         """
-        self._survey = self._create_str_param(min_val, max_val, eq)
+        self._survey = self._create_str_param(eq, min_val, max_val)
         return self
-    
+
     def filter_survey_id(self, eq: FLOAT_PARAM = None,
         min_val: OPT_FLOAT = None, max_val: OPT_FLOAT = None) -> 'Query':
         """Filter on unique numeric ID for the survey.
@@ -200,9 +200,9 @@ class Query:
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
         """
-        self._survey_id = self._create_float_param(min_val, max_val, eq)
+        self._survey_id = self._create_float_param(eq, min_val, max_val)
         return self
-    
+
     def filter_cruise(self, eq: FLOAT_PARAM = None, min_val: OPT_FLOAT = None,
         max_val: OPT_FLOAT = None) -> 'Query':
         """Filter on cruise ID.
@@ -222,9 +222,9 @@ class Query:
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
         """
-        self._cruise = self._create_float_param(min_val, max_val, eq)
+        self._cruise = self._create_float_param(eq, min_val, max_val)
         return self
-    
+
     def filter_haul(self, eq: FLOAT_PARAM = None, min_val: OPT_FLOAT = None,
         max_val: OPT_FLOAT = None) -> 'Query':
         """Filter on haul identifier.
@@ -244,9 +244,9 @@ class Query:
                 maximum value filter should be applied. Defaults to None. Error
[…]
