+++ .github/workflows/docs.yml
@@ -29,6 +29,8 @@ jobs:
         run: pip install -e .[dev]
       - name: Build API docs
         run: pdoc --docformat google ./afscgap -o website/devdocs/
+      - name: Update apt cache
+        run: sudo apt-get update
       - name: Build paper
         run: cd inst; bash preview_paper.sh
       - name: Upload website
