### 2023-04-17 #93 @sampottinger
Hello all! @ocefpaf, the builder interface was implemented in https://github.com/SchmidtDSE/afscgap/pull/52.
