+++ afscgap/model.py
@@ -175,7 +175,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_latitude_dd(self) -> float:
+    def get_latitude(self) -> float:
         """Get the field labeled as latitude_dd in the API.
 
         Returns:
@@ -183,7 +183,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_longitude_dd(self) -> float:
+    def get_longitude(self) -> float:
         """Get the field labeled as longitude_dd in the API.
 
         Returns:
@@ -227,7 +227,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_cpue_kgha_maybe(self) -> OPT_FLOAT:
+    def get_cpue_weight_maybe(self, units='kg/ha') -> OPT_FLOAT:
         """Get the field labeled as cpue_kgha in the API.
 
         Returns:
@@ -237,53 +237,13 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_cpue_kgkm2_maybe(self) -> OPT_FLOAT:
-        """Get the field labeled as cpue_kgkm2 in the API.
+    def get_cpue_count_maybe(self, units='count/ha') -> OPT_FLOAT:
+        """Get the field labeled as cpue_kgha in the API.
 
         Returns:
-            Catch weight divided by net area (kg / km^2) if available. See
+            Catch weight divided by net area (kg / hectares) if available. See
             metadata. None if could not interpret as a float. If an inferred
             zero catch record, will be zero.
-        """
-        raise NotImplementedError('Use implementor.')
-
-    def get_cpue_kg1000km2_maybe(self) -> OPT_FLOAT:
-        """Get the field labeled as cpue_kg1000km2 in the API.
-
-        Returns:
-            Catch weight divided by net area (kg / km^2 * 1000) if available.
-            See metadata. None if could not interpret as a float. If an inferred
-            zero catch record, will be zero.
-        """
-        raise NotImplementedError('Use implementor.')
-
-    def get_cpue_noha_maybe(self) -> OPT_FLOAT:
-        """Get the field labeled as cpue_noha in the API.
-
-        Returns:
-            Catch number divided by net sweep area if available (count /
-            hectares). See metadata. None if could not interpret as a float. If
-            an inferred zero catch record, will be zero.
-        """
-        raise NotImplementedError('Use implementor.')
-
-    def get_cpue_nokm2_maybe(self) -> OPT_FLOAT:
-        """Get the field labeled as cpue_nokm2 in the API.
-
-        Returns:
-            Catch number divided by net sweep area if available (count / km^2).
-            See metadata. None if could not interpret as a float. If an inferred
-            zero catch record, will be zero.
-        """
-        raise NotImplementedError('Use implementor.')
-
-    def get_cpue_no1000km2_maybe(self) -> OPT_FLOAT:
-        """Get the field labeled as cpue_no1000km2 in the API.
-
-        Returns:
-            Catch number divided by net sweep area if available (count / km^2 *
-            1000). See metadata. None if could not interpret as a float. If an
-            inferred zero catch record, will be zero.
         """
         raise NotImplementedError('Use implementor.')
 
@@ -307,7 +267,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_bottom_temperature_c_maybe(self) -> OPT_FLOAT:
+    def get_bottom_temperature_maybe(self, units='c') -> OPT_FLOAT:
         """Get the field labeled as bottom_temperature_c in the API.
 
         Returns:
[…]
