### 2023-04-17 0af306a
Another draft of new interface.
