+++ afscgap/__init__.py
@@ -133,6 +133,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._year = self._create_float_param(eq, min_val, max_val)
         return self
@@ -158,6 +160,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._srvy = self._create_str_param(eq, min_val, max_val)
         return self
@@ -180,6 +184,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._survey = self._create_str_param(eq, min_val, max_val)
         return self
@@ -199,6 +205,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._survey_id = self._create_float_param(eq, min_val, max_val)
         return self
@@ -221,6 +229,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._cruise = self._create_float_param(eq, min_val, max_val)
         return self
@@ -243,6 +253,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._haul = self._create_float_param(eq, min_val, max_val)
         return self
@@ -262,6 +274,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._stratum = self._create_float_param(eq, min_val, max_val)
         return self
@@ -281,6 +295,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._station = self._create_str_param(eq, min_val, max_val)
         return self
@@ -300,6 +316,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._vessel_name = self._create_str_param(eq, min_val, max_val)
         return self
@@ -319,6 +337,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        Returns:
+            This object for chaining if desired.
         """
         self._vessel_id = self._create_float_param(eq, min_val, max_val)
         return self
@@ -343,6 +363,8 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
[…]
