+++ afscgap/model.py
@@ -175,17 +175,23 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_latitude(self) -> float:
+    def get_latitude(self, units='dd') -> float:
         """Get the field labeled as latitude_dd in the API.
 
+        Args:
+            units: The units to return this value in. Only supported is dd for
+                degrees. Deafults to dd.
         Returns:
             Latitude in decimal degrees associated with the haul.
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_longitude(self) -> float:
+    def get_longitude(self, units='dd') -> float:
         """Get the field labeled as longitude_dd in the API.
 
+        Args:
+            units: The units to return this value in. Only supported is dd for
+                degrees. Deafults to dd.
         Returns:
             Longitude in decimal degrees associated with the haul.
         """
@@ -228,30 +234,45 @@ class Record(HaulKeyable):
         raise NotImplementedError('Use implementor.')
 
     def get_cpue_weight_maybe(self, units='kg/ha') -> OPT_FLOAT:
-        """Get the field labeled as cpue_kgha in the API.
+        """Get a field labeled as cpue_* in the API.
 
+        Args:
+            units: The desired units for the catch per unit effort. Options:
+                kg/ha, kg/km2, kg1000/km2. Defaults to kg/ha.
+        
         Returns:
-            Catch weight divided by net area (kg / hectares) if available. See
+            Catch weight divided by net area (in given units) if available. See
             metadata. None if could not interpret as a float. If an inferred
             zero catch record, will be zero.
         """
         raise NotImplementedError('Use implementor.')
 
     def get_cpue_count_maybe(self, units='count/ha') -> OPT_FLOAT:
-        """Get the field labeled as cpue_kgha in the API.
+        """Get the field labeled as cpue_* in the API.
 
+        Get the catch per unit effort from the record with one of the following
+        units: kg/ha, kg/km2, kg1000/km2.
+
+        Args:
+            units: The desired units for the catch per unit effort. Options:
+                count/ha, count/km2, and count1000/km2. Defaults to count/ha.
+        
         Returns:
-            Catch weight divided by net area (kg / hectares) if available. See
+            Catch weight divided by net area (in given units) if available. See
             metadata. None if could not interpret as a float. If an inferred
             zero catch record, will be zero.
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_weight_kg_maybe(self) -> OPT_FLOAT:
+    def get_weight_maybe(self, units='kg') -> OPT_FLOAT:
         """Get the field labeled as weight_kg in the API.
 
+        Args:
+            units: The units in which the weight should be returned. Options are
+                g, kg for grams and kilograms respectively. Deafults to kg.
+        
         Returns:
-            Taxon weight (kg) if available. See metadata. None if could not
+            Taxon weight if available. See metadata. None if could not
             interpret as a float. If an inferred zero catch record, will be
             zero.
         """
@@ -270,84 +291,125 @@ class Record(HaulKeyable):
     def get_bottom_temperature_maybe(self, units='c') -> OPT_FLOAT:
         """Get the field labeled as bottom_temperature_c in the API.
 
+        Args:
+            units: The units in which the temperature should be returned.
+                Options: c or f for Celcius and Fahrenheit respectively.
+                Defaults to c.
+        
         Returns:
             Bottom temperature associated with observation / inferrence if
-            available in Celsius. None if not given or could not interpret as a
-            float.
+            available in desired units. None if not given or could not interpret
+            as a float.
         """
         raise NotImplementedError('Use implementor.')
 
     def get_surface_temperature_maybe(self, units='c') -> OPT_FLOAT:
[…]
