+++ afscgap/__init__.py
@@ -133,6 +133,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+
         Returns:
             This object for chaining if desired.
         """
@@ -160,6 +161,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -184,6 +186,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -205,6 +208,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -229,6 +233,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -253,6 +258,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -274,6 +280,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -295,6 +302,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -316,6 +324,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -337,6 +346,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -363,6 +373,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -384,6 +395,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -405,6 +417,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+        
         Returns:
             This object for chaining if desired.
         """
@@ -426,6 +439,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
[…]
