### 2023-04-17 c44b06f
Through refactor of ApiRecord for #55
