+++ afscgap/model.py
@@ -175,23 +175,25 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_latitude(self, units='dd') -> float:
+    def get_latitude(self, units: str = 'dd') -> float:
         """Get the field labeled as latitude_dd in the API.
 
         Args:
             units: The units to return this value in. Only supported is dd for
                 degrees. Deafults to dd.
+        
         Returns:
             Latitude in decimal degrees associated with the haul.
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_longitude(self, units='dd') -> float:
+    def get_longitude(self, units: str = 'dd') -> float:
         """Get the field labeled as longitude_dd in the API.
 
         Args:
             units: The units to return this value in. Only supported is dd for
                 degrees. Deafults to dd.
+        
         Returns:
             Longitude in decimal degrees associated with the haul.
         """
@@ -233,7 +235,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_cpue_weight_maybe(self, units='kg/ha') -> OPT_FLOAT:
+    def get_cpue_weight_maybe(self, units: str = 'kg/ha') -> OPT_FLOAT:
         """Get a field labeled as cpue_* in the API.
 
         Args:
@@ -247,7 +249,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_cpue_count_maybe(self, units='count/ha') -> OPT_FLOAT:
+    def get_cpue_count_maybe(self, units: str = 'count/ha') -> OPT_FLOAT:
         """Get the field labeled as cpue_* in the API.
 
         Get the catch per unit effort from the record with one of the following
@@ -264,7 +266,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_weight_maybe(self, units='kg') -> OPT_FLOAT:
+    def get_weight_maybe(self, units: str = 'kg') -> OPT_FLOAT:
         """Get the field labeled as weight_kg in the API.
 
         Args:
@@ -288,7 +290,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_bottom_temperature_maybe(self, units='c') -> OPT_FLOAT:
+    def get_bottom_temperature_maybe(self, units: str = 'c') -> OPT_FLOAT:
         """Get the field labeled as bottom_temperature_c in the API.
 
         Args:
@@ -303,7 +305,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_surface_temperature_maybe(self, units='c') -> OPT_FLOAT:
+    def get_surface_temperature_maybe(self, units: str = 'c') -> OPT_FLOAT:
         """Get the field labeled as surface_temperature_c in the API.
 
         Args:
@@ -317,7 +319,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_depth(self, units='m') -> float:
+    def get_depth(self, units: str = 'm') -> float:
         """Get the field labeled as depth_m in the API.
 
         Args:
@@ -329,7 +331,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
-    def get_distance_fished(self, units='km') -> float:
+    def get_distance_fished(self, units: str = 'km') -> float:
         """Get the field labeled as distance_fished_km in the API.
 
         Args:
@@ -341,7 +343,7 @@ class Record(HaulKeyable):
         """
         raise NotImplementedError('Use implementor.')
 
[…]
