+++ afscgap/client.py
@@ -555,21 +555,29 @@ class ApiRecord(afscgap.model.Record):
         """
         return self._date_time
 
-    def get_latitude_dd(self) -> float:
+    def get_latitude(self, units: str = 'dd') -> float:
         """Get the field labeled as latitude_dd in the API.
 
+        Args:
+            units: The units to return this value in. Only supported is dd for
+                degrees. Deafults to dd.
+        
         Returns:
             Latitude in decimal degrees associated with the haul.
         """
-        return self._latitude_dd
+        return convert.convert_degrees(self._latitude_dd, units)
 
-    def get_longitude_dd(self) -> float:
+    def get_longitude(self, units: str = 'dd') -> float:
         """Get the field labeled as longitude_dd in the API.
 
+        Args:
+            units: The units to return this value in. Only supported is dd for
+                degrees. Deafults to dd.
+        
         Returns:
             Longitude in decimal degrees associated with the haul.
         """
-        return self._longitude_dd
+        return convert.convert_degrees(self._longitude_dd, units)
 
     def get_species_code(self) -> float:
         """Get the field labeled as species_code in the API.
@@ -606,68 +614,58 @@ class ApiRecord(afscgap.model.Record):
         """
         return self._taxon_confidence
 
-    def get_cpue_kgha_maybe(self) -> OPT_FLOAT:
-        """Get the field labeled as cpue_kgha in the API.
+    def get_cpue_weight_maybe(self, units: str = 'kg/ha') -> OPT_FLOAT:
+        """Get a field labeled as cpue_* in the API.
 
+        Args:
+            units: The desired units for the catch per unit effort. Options:
+                kg/ha, kg/km2, kg1000/km2. Defaults to kg/ha.
+        
         Returns:
-            Catch weight divided by net area (kg / hectares) if available. See
-            metadata. None if could not interpret as a float.
+            Catch weight divided by net area (in given units) if available. See
+            metadata. None if could not interpret as a float. If an inferred
+            zero catch record, will be zero.
         """
-        return self._cpue_kgha
+        return {
+            'kg/ha': self._cpue_kgha,
+            'kg/km2': self._cpue_kgkm2,
+            'kg1000/km2': self._cpue_kg1000km2
+        }[units.lower()]
 
-    def get_cpue_kgkm2_maybe(self) -> OPT_FLOAT:
-        """Get the field labeled as cpue_kgkm2 in the API.
+    def get_cpue_count_maybe(self, units: str = 'kg/ha') -> OPT_FLOAT:
+        """Get the field labeled as cpue_* in the API.
 
+        Get the catch per unit effort from the record with one of the following
+        units: kg/ha, kg/km2, kg1000/km2.
+
+        Args:
+            units: The desired units for the catch per unit effort. Options:
+                count/ha, count/km2, and count1000/km2. Defaults to count/ha.
+        
         Returns:
-            Catch weight divided by net area (kg / km^2) if available. See
-            metadata. None if could not interpret as a float.
+            Catch weight divided by net area (in given units) if available. See
+            metadata. None if could not interpret as a float. If an inferred
+            zero catch record, will be zero.
         """
-        return self._cpue_kgkm2
+        return {
+            'count/ha': self._cpue_noha,
+            'count/km2': self._cpue_nokm2, 
+            'count1000/km2': self._cpue_no1000km2
+        }[units]
 
-    def get_cpue_kg1000km2_maybe(self) -> OPT_FLOAT:
-        """Get the field labeled as cpue_kg1000km2 in the API.
-
-        Returns:
-            Catch weight divided by net area (kg / km^2 * 1000) if available.
-            See metadata. None if could not interpret as a float.
-        """
[…]
