### 2023-04-17 f9a489f
Draft of migration for ZeroCatchHaulDecorator
