+++ afscgap/inference.py
@@ -604,21 +604,29 @@ class ZeroCatchHaulDecorator(afscgap.model.Record):
         """
         return self._haul.get_date_time()
 
-    def get_latitude_dd(self) -> float:
+    def get_latitude(self, units: str = 'dd') -> float:
         """Get the field labeled as latitude_dd in the API.
 
+        Args:
+            units: The units to return this value in. Only supported is dd for
+                degrees. Deafults to dd.
+        
         Returns:
             Latitude in decimal degrees associated with the haul.
         """
-        return self._haul.get_latitude_dd()
+        return convert.convert_degrees(self._haul.get_latitude_dd(), units)
 
-    def get_longitude_dd(self) -> float:
+    def get_longitude(self, units: str = 'dd') -> float:
         """Get the field labeled as longitude_dd in the API.
 
+        Args:
+            units: The units to return this value in. Only supported is dd for
+                degrees. Deafults to dd.
+        
         Returns:
             Longitude in decimal degrees associated with the haul.
         """
-        return self._haul.get_longitude_dd()
+        return convert.convert_degrees(self._haul.get_longitude_dd(), units)
 
     def get_species_code(self) -> float:
         """Get the field labeled as species_code in the API.
@@ -654,59 +662,48 @@ class ZeroCatchHaulDecorator(afscgap.model.Record):
         """
         return 'Unassessed'
 
-    def get_cpue_kgha_maybe(self) -> OPT_FLOAT:
-        """Get catch weight divided by net area (kg / hectares).
+    def get_cpue_weight_maybe(self, units: str = 'kg/ha') -> OPT_FLOAT:
+        """Get a field labeled as cpue_* in the API.
 
+        Args:
+            units: The desired units for the catch per unit effort. Options:
+                kg/ha, kg/km2, kg1000/km2. Defaults to kg/ha.
+        
         Returns:
-            Always returns 0.
+            Catch weight divided by net area (in given units) if available. See
+            metadata. None if could not interpret as a float. If an inferred
+            zero catch record, will be zero.
         """
         return 0
 
-    def get_cpue_kgkm2_maybe(self) -> OPT_FLOAT:
-        """Get catch weight divided by net area (kg / km^2).
+    def get_cpue_count_maybe(self, units: str = 'kg/ha') -> OPT_FLOAT:
+        """Get the field labeled as cpue_* in the API.
 
+        Get the catch per unit effort from the record with one of the following
+        units: kg/ha, kg/km2, kg1000/km2.
+
+        Args:
+            units: The desired units for the catch per unit effort. Options:
+                count/ha, count/km2, and count1000/km2. Defaults to count/ha.
+        
         Returns:
-            Always returns 0.
+            Catch weight divided by net area (in given units) if available. See
+            metadata. None if could not interpret as a float. If an inferred
+            zero catch record, will be zero.
         """
         return 0
 
-    def get_cpue_kg1000km2_maybe(self) -> OPT_FLOAT:
-        """Get catch weight divided by net area (kg / km^2 * 1000).
+    def get_weight_maybe(self, units='kg') -> OPT_FLOAT:
+        """Get the field labeled as weight_kg in the API.
 
+        Args:
+            units: The units in which the weight should be returned. Options are
+                g, kg for grams and kilograms respectively. Deafults to kg.
+        
         Returns:
-            Always returns 0.
-        """
-        return 0
-
-    def get_cpue_noha_maybe(self) -> OPT_FLOAT:
-        """Get catch number divided by net sweep area.
-
-        Returns:
-            Always returns 0.
-        """
-        return 0
-
-    def get_cpue_nokm2_maybe(self) -> OPT_FLOAT:
[…]
