+++ afscgap/__init__.py
@@ -161,7 +161,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -186,7 +186,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -208,7 +208,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -233,7 +233,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -258,7 +258,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -280,7 +280,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -302,7 +302,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -324,7 +324,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -346,7 +346,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -373,7 +373,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -395,7 +395,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -417,7 +417,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -439,7 +439,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
-        
+
         Returns:
             This object for chaining if desired.
         """
@@ -461,7 +461,7 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
[…]
