+++ afscgap/client.py
@@ -561,11 +561,13 @@ class ApiRecord(afscgap.model.Record):
         Args:
             units: The units to return this value in. Only supported is dd for
                 degrees. Deafults to dd.
-        
+
         Returns:
             Latitude in decimal degrees associated with the haul.
         """
-        return convert.convert_degrees(self._latitude_dd, units)
+        return afscgap.model.assert_float_present(
+            afscgap.convert.convert_degrees(self._latitude_dd, units)
+        )
 
     def get_longitude(self, units: str = 'dd') -> float:
         """Get the field labeled as longitude_dd in the API.
@@ -573,11 +575,13 @@ class ApiRecord(afscgap.model.Record):
         Args:
             units: The units to return this value in. Only supported is dd for
                 degrees. Deafults to dd.
-        
+
         Returns:
             Longitude in decimal degrees associated with the haul.
         """
-        return convert.convert_degrees(self._longitude_dd, units)
+        return afscgap.model.assert_float_present(
+            afscgap.convert.convert_degrees(self._longitude_dd, units)
+        )
 
     def get_species_code(self) -> float:
         """Get the field labeled as species_code in the API.
@@ -620,7 +624,7 @@ class ApiRecord(afscgap.model.Record):
         Args:
             units: The desired units for the catch per unit effort. Options:
                 kg/ha, kg/km2, kg1000/km2. Defaults to kg/ha.
-        
+
         Returns:
             Catch weight divided by net area (in given units) if available. See
             metadata. None if could not interpret as a float. If an inferred
@@ -641,7 +645,7 @@ class ApiRecord(afscgap.model.Record):
         Args:
             units: The desired units for the catch per unit effort. Options:
                 count/ha, count/km2, and count1000/km2. Defaults to count/ha.
-        
+
         Returns:
             Catch weight divided by net area (in given units) if available. See
             metadata. None if could not interpret as a float. If an inferred
@@ -649,7 +653,7 @@ class ApiRecord(afscgap.model.Record):
         """
         return {
             'count/ha': self._cpue_noha,
-            'count/km2': self._cpue_nokm2, 
+            'count/km2': self._cpue_nokm2,
             'count1000/km2': self._cpue_no1000km2
         }[units]
 
@@ -659,13 +663,13 @@ class ApiRecord(afscgap.model.Record):
         Args:
             units: The units in which the weight should be returned. Options are
                 g, kg for grams and kilograms respectively. Deafults to kg.
-        
+
         Returns:
             Taxon weight if available. See metadata. None if could not
             interpret as a float. If an inferred zero catch record, will be
             zero.
         """
-        return convert.convert_weight(self._weight_kg, units)
+        return afscgap.convert.convert_weight(self._weight_kg, units)
 
     def get_count_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as count in the API.
@@ -683,13 +687,16 @@ class ApiRecord(afscgap.model.Record):
             units: The units in which the temperature should be returned.
                 Options: c or f for Celcius and Fahrenheit respectively.
                 Defaults to c.
-        
+
         Returns:
             Bottom temperature associated with observation / inferrence if
             available in desired units. None if not given or could not interpret
             as a float.
         """
-        return convert.convert_temperature(self._bottom_temperature_c, units)
+        return afscgap.convert.convert_temperature(
+            self._bottom_temperature_c,
+            units
+        )
 
     def get_surface_temperature_maybe(self, units: str = 'c') -> OPT_FLOAT:
         """Get the field labeled as surface_temperature_c in the API.
@@ -698,12 +705,15 @@ class ApiRecord(afscgap.model.Record):
             units: The units in which the temperature should be returned.
[…]
