+++ afscgap/__init__.py
@@ -358,9 +358,10 @@ class Query:
         """Filter on the date and time of the haul.
 
         Filter on the date and time of the haul as an ISO 8601 string. If given
-        an ISO 8601 string, will convert from ISO 8601 to the API datetime
-        string format. Similarly, if given a dictionary, all values matching an
-        ISO 8601 string will be converted to the API datetime string format.
+        an ISO 8601 string, will afscgap.convert from ISO 8601 to the API
+        datetime string format. Similarly, if given a dictionary, all values
+        matching an ISO 8601 string will be afscgap.converted to the API
+        datetime string format.
 
         Args:
             eq: The exact value that must be matched for a record to be
@@ -380,8 +381,9 @@ class Query:
         self._date_time = self._create_str_param(eq, min_val, max_val)
         return self
 
-    def filter_latitude_dd(self, eq: FLOAT_PARAM = None,
-        min_val: OPT_FLOAT = None, max_val: OPT_FLOAT = None) -> 'Query':
+    def filter_latitude(self, eq: FLOAT_PARAM = None,
+        min_val: OPT_FLOAT = None, max_val: OPT_FLOAT = None,
+        units: str = 'dd') -> 'Query':
         """Filter on latitude in decimal degrees associated with the haul.
 
         Args:
@@ -395,15 +397,23 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+            units: The units in which the filter values are provided. Currently
+                only dd supported. Ignored if given eq value containing ORDS
+                query.
 
         Returns:
             This object for chaining if desired.
         """
-        self._latitude_dd = self._create_float_param(eq, min_val, max_val)
+        self._latitude_dd = self._create_float_param(
+            afscgap.convert.unconvert_degrees(eq, units),
+            afscgap.convert.unconvert_degrees(min_val, units),
+            afscgap.convert.unconvert_degrees(max_val, units)
+        )
         return self
 
-    def filter_longitude_dd(self, eq: FLOAT_PARAM = None,
-        min_val: OPT_FLOAT = None, max_val: OPT_FLOAT = None) -> 'Query':
+    def filter_longitude(self, eq: FLOAT_PARAM = None,
+        min_val: OPT_FLOAT = None, max_val: OPT_FLOAT = None,
+        units: str = 'dd') -> 'Query':
         """Filter on longitude in decimal degrees associated with the haul.
 
         Args:
@@ -417,11 +427,17 @@ class Query:
             max_val: The maximum allowed value, inclusive. Pass None if no
                 maximum value filter should be applied. Defaults to None. Error
                 thrown if eq also proivded.
+            units: The units in which the filter values are provided. Currently
+                only dd supported.
 
         Returns:
             This object for chaining if desired.
         """
-        self._longitude_dd = self._create_float_param(eq, min_val, max_val)
+        self._longitude_dd = self._create_float_param(
+            afscgap.convert.unconvert_degrees(eq, units),
+            afscgap.convert.unconvert_degrees(min_val, units),
+            afscgap.convert.unconvert_degrees(max_val, units)
+        )
         return self
 
     def filter_species_code(self, eq: FLOAT_PARAM = None,
@@ -512,61 +528,12 @@ class Query:
         self._taxon_confidence = self._create_str_param(eq, min_val, max_val)
         return self
 
-    def filter_cpue_kgha(self, eq: FLOAT_PARAM = None,
-        min_val: OPT_FLOAT = None, max_val: OPT_FLOAT = None) -> 'Query':
[…]
