+++ afscgap/convert.py
@@ -9,6 +9,7 @@ LICENSE.md.
 """
 import re
 
+from afscgap.typesdef import FLOAT_PARAM
 from afscgap.typesdef import OPT_FLOAT
 from afscgap.typesdef import STR_PARAM
 
@@ -27,14 +28,30 @@ AREA_CONVERTERS = {
     'km2': lambda x: x * 0.01
 }
 
+AREA_UNCONVERTERS = {
+    'ha': lambda x: x,
+    'm2': lambda x: x / 10000,
+    'km2': lambda x: x / 0.01
+}
+
 DISTANCE_CONVERTERS = {
     'm': lambda x: x,
     'km': lambda x: x / 1000
 }
 
+DISTANCE_UNCONVERTERS = {
+    'm': lambda x: x,
+    'km': lambda x: x * 1000
+}
+
 TEMPERATURE_CONVERTERS = {
     'c': lambda x: x,
     'f': lambda x: x * 9 / 5 + 32
+}
+
+TEMPERATURE_UNCONVERTERS = {
+    'c': lambda x: x,
+    'f': lambda x: (x - 32) * 5 / 9
 }
 
 TIME_CONVERTERS = {
@@ -43,8 +60,19 @@ TIME_CONVERTERS = {
     'min': lambda x: x * 60
 }
 
+TIME_UNCONVERTERS = {
+    'day': lambda x: x * 24,
+    'hr': lambda x: x,
+    'min': lambda x: x / 60
+}
+
 WEIGHT_CONVERTERS = {
     'g': lambda x: x * 1000,
+    'kg': lambda x: x
+}
+
+WEIGHT_UNCONVERTERS = {
+    'g': lambda x: x / 1000,
     'kg': lambda x: x
 }
 
@@ -158,12 +186,46 @@ def convert_area(target: OPT_FLOAT, units: str) -> OPT_FLOAT:
     return AREA_CONVERTERS[units](target)
 
 
+def unconvert_area(target: FLOAT_PARAM, units: str) -> FLOAT_PARAM:
+    """Standardize an area to the API-native units (hectare).
+
+    Args:
+        target: The value to convert in hectares.
+        units: The units of value.
+
+    Returns:
+        The converted value. Note that, if target is None, will return None.
+    """
+    if target is None:
+        return None
+
+    if isinstance(target, dict):
+        return target
+
+    return AREA_UNCONVERTERS[units](target)
+
+
 def convert_degrees(target: OPT_FLOAT, units: str) -> OPT_FLOAT:
     """Convert targets from degrees to another units.
 
     Args:
         target: The value to convert which may be None.
         units: Desired units.
+
+    Returns:
+        The same value input after asserting that units are dd, the only
+        supported units.
+    """
+    assert units == 'dd'
+    return target
+
+
+def unconvert_degrees(target: FLOAT_PARAM, units: str) -> FLOAT_PARAM:
+    """Standardize a degree to the API-native units (degrees).
+
+    Args:
+        target: The value to convert which may be None.
+        units: The units of value.
 
     Returns:
         The same value input after asserting that units are dd, the only
@@ -183,7 +245,29 @@ def convert_distance(target: OPT_FLOAT, units: str) -> OPT_FLOAT:
     Returns:
         The converted value. Note that, if target is None, will return None.
     """
+    if target is None:
+        return None
+
     return DISTANCE_CONVERTERS[units](target)
+
+
+def unconvert_distance(target: FLOAT_PARAM, units: str) -> FLOAT_PARAM:
+    """Convert a linear distance to the API-native units (meters).
+
+    Args:
+        target: The value to convert in meters.
+        units: The units of value.
+
+    Returns:
[…]
