### 2023-04-18 #93 @sampottinger
> Mind renaming to something like get_bottom_temperature(unit="celsius"|'c')
Thanks @ayushanand18! I have implemented in https://github.com/SchmidtDSE/afscgap/pull/55
