### 2023-04-18 #93 @sampottinger
I am confirming https://github.com/SchmidtDSE/afscgap/pull/54 which includes the following requested changes before they are resolved on main:

- https://github.com/SchmidtDSE/afscgap/issues/50
- https://github.com/SchmidtDSE/afscgap/issues/51

https://github.com/SchmidtDSE/afscgap/pull/54 will also soon include updates to documentation.
