### 2023-04-18 f498731
Make query setters non-keyword
