+++ afscgap/__init__.py
@@ -122,6 +122,10 @@ class Query:
         max_val: OPT_FLOAT = None) -> 'Query':
         """Filter on year for the survey in which this observation was made.
 
+        Filter on year for the survey in which this observation was made,
+        ovewritting all prior year filters on this Query if one was previously
+        set.
+
         Args:
             eq: The exact value that must be matched for a record to be
                 returned. Pass None if no equality filter should be applied.
@@ -148,7 +152,7 @@ class Query:
         made. Pass None if no filter should be applied. Defaults to None. Note
         that common values include: NBS (N Bearing Sea), EBS (SE Bearing Sea),
         BSS (Bearing Sea Slope), GOA (Gulf of Alaska), and AI (Aleutian
-        Islands).
+        Islands). Overwrites all prior srvy filters if set on this Query.
 
         Args:
             eq: The exact value that must be matched for a record to be
@@ -173,7 +177,7 @@ class Query:
         """Filter on survey long name.
 
         Filter on long form description of the survey in which the observation
-        was made.
+        was made. Overwrites all prior survey filters if set on this Query.
 
         Args:
             eq: The exact value that must be matched for a record to be
@@ -196,6 +200,9 @@ class Query:
     def filter_survey_id(self, eq: FLOAT_PARAM = None,
         min_val: OPT_FLOAT = None, max_val: OPT_FLOAT = None) -> 'Query':
         """Filter on unique numeric ID for the survey.
+
+        Filter on unique numeric ID for the survey, overwritting prior survey ID
+        filters if set on this Query.
 
         Args:
             eq: The exact value that must be matched for a record to be
@@ -220,7 +227,7 @@ class Query:
         """Filter on cruise ID.
 
         Filter on an ID uniquely identifying the cruise in which the observation
-        was made.
+        was made. Overwrites all prior cruise filters on this Query.
 
         Args:
             eq: The exact value that must be matched for a record to be
@@ -245,7 +252,7 @@ class Query:
         """Filter on haul identifier.
 
         Filter on an ID uniquely identifying the haul in which this observation
-        was made.
+        was made. Overwrites all prior haul filters on this Query.
 
         Args:
             eq: The exact value that must be matched for a record to be
@@ -269,6 +276,9 @@ class Query:
         max_val: OPT_FLOAT = None) -> 'Query':
         """Filter on unique ID for statistical area / survey combination.
 
+        Filter on unique ID for statistical area / survey combination,
+        overwritting all prior stratum filters on Query.
+
         Args:
             eq: The exact value that must be matched for a record to be
                 returned. Pass None if no equality filter should be applied.
@@ -290,6 +300,9 @@ class Query:
     def filter_station(self, eq: STR_PARAM = None, min_val: OPT_STR = None,
         max_val: OPT_STR = None) -> 'Query':
         """Filter on station associated with the survey.
+
+        Filter on station associated with the survey, overwritting all prior
+        station filters on this Query.
 
         Args:
             eq: The exact value that must be matched for a record to be
@@ -313,6 +326,9 @@ class Query:
         min_val: OPT_STR = None, max_val: OPT_STR = None) -> 'Query':
         """Filter on unique ID describing the vessel that made this observation.
 
+        Filter on unique ID describing the vessel that made this observation,
+        overwritting all prior vessel name filters on this Query.
+
         Args:
             eq: The exact value that must be matched for a record to be
                 returned. Pass None if no equality filter should be applied.
@@ -334,6 +350,9 @@ class Query:
     def filter_vessel_id(self, eq: FLOAT_PARAM = None,
         min_val: OPT_FLOAT = None, max_val: OPT_FLOAT = None) -> 'Query':
[…]
