+++ README.md
@@ -81,40 +81,48 @@ Visualization tools are available to help both programmers and non-programmers s
 <br>
 
 #### Basic queries
-The `afscgap.query` method is the main entry point into Python-based utilization. Calls can be written manually or generated in the [visual analytics tool](https://app.pyafscgap.org). For example, this requests all records of Pasiphaea pacifica in 2021 from the Gulf of Alaska to get the median bottom temperature when they were observed:
+The `afscgap.Query` object is the main entry point into Python-based utilization. Calls can be written manually or generated in the [visual analytics tool](https://app.pyafscgap.org). For example, this requests all records of Pasiphaea pacifica in 2021 from the Gulf of Alaska to get the median bottom temperature when they were observed:
 
 ```
 import statistics
 
 import afscgap
 
-results = afscgap.query(
-    year=2021,
-    srvy='GOA',
-    scientific_name='Pasiphaea pacifica'
-)
+# Build query
+query = afscgap.Query()
+query.filter_year(eq=2021)
+query.filter_srvy(eq='GOA')
+query.filter_scientific_name(eq='Pasiphaea pacifica')
+results = query.execute()
 
-temperatures = [record.get_bottom_temperature_c() for record in results]
+# Get temperatures in Celsius
+temperatures = [record.get_bottom_temperature(units='c') for record in results]
+
+# Take the median
 print(statistics.median(temperatures))
 ```
 
-Note that `afscgap.query` returns a [Cursor](https://pyafscgap.org/devdocs/afscgap/cursor.html#Cursor). One can iterate over this `Cursor` to access [Record](https://pyafscgap.org/devdocs/afscgap/model.html#Record) objects. You can do this with list comprehensions, maps, etc or with a good old for loop like in this example which gets a histogram of haul temperatures:
+Note that `afscgap.Query.execute` returns a [Cursor](https://pyafscgap.org/devdocs/afscgap/cursor.html#Cursor). One can iterate over this `Cursor` to access [Record](https://pyafscgap.org/devdocs/afscgap/model.html#Record) objects. You can do this with list comprehensions, maps, etc or with a good old for loop like in this example which gets a histogram of haul temperatures:
 
 ```
+# Mapping from temperature in Celsius to count
 count_by_temperature_c = {}
 
-results = afscgap.query(
-    year=2021,
-    srvy='GOA',
-    scientific_name='Pasiphaea pacifica'
-)
+# Build query
+query = afscgap.Query()
+query.filter_year(eq=2021)
+query.filter_srvy(eq='GOA')
+query.filter_scientific_name(eq='Pasiphaea pacifica')
+results = query.execute()
 
+# Iterate through results and count
 for record in results:
-    temp = record.get_bottom_temperature_c()
+    temp = record.get_bottom_temperature(units='c')
     temp_rounded = round(temp)
     count = count_by_temperature_c.get(temp_rounded, 0) + 1
     count_by_temperature_c[temp_rounded] = count
 
+# Print the result
 print(count_by_temperature_c)
 ```
 
@@ -125,24 +133,24 @@ See [data structure section](#data-structure). Using an iterator will have the l
 #### Enable absence data
 One of the major limitations of the official API is that it only provides presence data. However, this library can optionally infer absence or "zero catch" records using a separate static file produced by NOAA AFSC GAP. The [algorithm and details for absence inference](#absence-vs-presence-data) is further discussed below.
 
-Absence data / "zero catch" records inference can be turned on by setting `presence_only` to false in `query`. To demonstrate, this example finds total area swept and total weight for Gadus macrocephalus from the Aleutian Islands in 2021:
[…]
