+++ README.md
@@ -286,7 +286,7 @@ query.filter_scientific_name(eq='Pasiphaea pacifica')
 results = query.execute()
 
 # Sum weight
-weights = map(lambda x: x.get_weight(units='kg'))
+weights = map(lambda x: x.get_weight(units='kg'), results)
 total_weight = sum(weights)
 print(total_weight)
 ```
@@ -304,7 +304,7 @@ query.filter_scientific_name(eq='Pasiphaea pacifica')
 results = query.execute()
 
 # Sum weight
-weights = map(lambda x: x.get_weight(units='kg'))
+weights = map(lambda x: x.get_weight(units='kg'), results)
 total_weight = sum(weights)
 print(total_weight)
 ```
@@ -322,7 +322,7 @@ query.filter_scientific_name(eq='Pasiphaea pacifica')
 results = query.execute()
 
 # Sum weight
-weights = map(lambda x: x.get_weight(units='kg'))
+weights = map(lambda x: x.get_weight(units='kg'), results)
 total_weight = sum(weights)
 print(total_weight)
 ```
@@ -341,7 +341,7 @@ import afscgap
 query = afscgap.Query()
 query.filter_year(eq=2021)
 query.filter_latitude({'$between': [56, 57]})
-query.filter_latitude({'$between': [-161, -160]})
+query.filter_longitude({'$between': [-161, -160]})
 results = query.execute()
 
 # Summarize
@@ -354,7 +354,7 @@ for record in results:
     count_by_common_name[common_name] = count
 
 # Print
-print(count_by_common_name['Pacific cod'])
+print(count_by_common_name['walleye pollock'])
 ```
 
 For more info about the options available, consider the [Oracle docs](https://docs.oracle.com/en/database/oracle/oracle-rest-data-services/19.2/aelig/developing-REST-applications.html#GUID-F0A4D4F9-443B-4EB9-A1D3-1CDE0A8BAFF2) or a helpful unaffiliated [getting started tutorial from Jeff Smith](https://www.thatjeffsmith.com/archive/2019/09/some-query-filtering-examples-in-ords/).
@@ -365,6 +365,8 @@ For more info about the options available, consider the [Oracle docs](https://do
 By default, the library will iterate through all results and handle pagination behind the scenes. However, one can also request an individual page:
 
 ```
+import afscgap 
+
 query = afscgap.Query()
 query.filter_year(eq=2021)
 query.filter_srvy(eq='GOA')
@@ -378,12 +380,15 @@ print(len(results_for_page))  # Will print 32 (results contains 52 records)
 Client code can also change the pagination behavior used when iterating:
 
 ```
+import afscgap
+
 query = afscgap.Query()
 query.filter_year(eq=2021)
 query.filter_srvy(eq='GOA')
 query.filter_scientific_name(eq='Gadus macrocephalus')
 query.set_start_offset(10)
 query.set_limit(200)
+query.set_filter_incomplete(True)
 results = query.execute()
 
 for record in results:
@@ -516,14 +521,14 @@ query.set_presence_only(False)
 results = query.execute()
 
 def simplify_record(full_record):
-    latitude = full_record.get_latitude_dd()
-    longitude = full_record.get_longitude_dd()
+    latitude = full_record.get_latitude(units='dd')
+    longitude = full_record.get_longitude(units='dd')
     geohash = geolib.geohash.encode(latitude, longitude, 5)
     
     return {
         'geohash': geohash,
-        'area': full_record.get_area_swept_ha(),
-        'weight': full_record.get_weight_kg()
+        'area': full_record.get_area_swept(units='ha'),
+        'weight': full_record.get_weight(units='kg')
     }
 
 def combine_record(a, b):
@@ -622,6 +627,8 @@ Metadata fields such as `year` are always required to make a `Record` whereas ot
[…]
