### 2023-04-18 #93 @ayushanand18
> I've started that at [SchmidtDSE/afscgap#51](https://github.com/SchmidtDSE/afscgap/issues/51)

That looks great!

> Thanks for this suggestion but I may pass on this if that's ok .

It's okay. At the end of the day, we need to follow what's best for the user of the package.

> To that end, there is a `to_dict` method on `afscgap.model.Record` for those that prefer that key-based approach or want to read into something like `pandas.DataFrame` [...]

Sounds good. But that leaves me a question of whether the `afsc.model.Record` object returns a nested dictionary in any case with the `to_dict()` method. If the answer is yes, then probably we might include a function `to_pandas()` that unnests the nested dictionary and returns a `DataFrame` object so that user doesn't take the pain. But again this is just a suggestion.
