### 2023-04-18 SchmidtDSE/afscgap#50 @ocefpaf
I really like the new syntax:

```yaml
        query = afscgap.Query(requestor=self._mock_requestor)
        query.filter_year(eq=2021)
        query.filter_latitude(eq={'$gte': 56.99, '$lte': 57.04})
        query.filter_longitude(eq={'$gte': -143.96, '$lte': -144.01})
        query.set_filter_incomplete(True)
        results = list(query.execute())
```
