### 2023-04-18 #93 @sampottinger
> I hear you. My day job is to try to make that better and, most of the times, we fail :-/

@ocefpaf - I'm sorry I didn't mean it like that! Perhaps just that the docstrings are extra convenient in this context :sweat_smile: . The upstream documentation for the datasets have to address both internal and external users with different kinds of access permissions / modalities. It just means it's a big schema with lots of (great) data but also takes a little while to get oriented towards. I think / hope that the docstrings help a bit but this library also has a simpler task since it knows which dataset the user is working with. I can also do a few things to help the user since I know they are in Python that the upstream service can't do. Also, thanks for your comment on the issue in the repo!

> That looks great!

Thank you! Appreciate your suggestions @ayushanand18.

> object returns a nested dictionary in any case with the to_dict() method

It is not nested! It uses the same field names and linear record structure from the API service.
