### 2023-04-26 e583cb4
Cleanup the quickstart section.
