+++ README.md
@@ -15,7 +15,24 @@ See [webpage](https://pyafscgap.org), [project Github](https://github.com/Schmid
 <br>
 
 ## Quickstart
-You don't need any local software to get started! To learn about the dataset, explore a visual analytics app at [https://app.pyafscgap.org](https://app.pyafscgap.org) (no code needed!). When ready, learn how to develop with these tools in a free [hosted notebook tutorial](https://mybinder.org/v2/gh/SchmidtDSE/afscgap/main?urlpath=/tree/index.ipynb).
+Taking your first step is easy!
+
+**Explore the data in a UI:** To learn about the dataset, explore a visual analytics app at [https://app.pyafscgap.org](https://app.pyafscgap.org).
+
+**Try out a tutorial in your browser:** Take a look at a free [hosted notebook tutorial](https://mybinder.org/v2/gh/SchmidtDSE/afscgap/main?urlpath=/tree/index.ipynb) with a practical analysis task all without installing any local software.
+
+**Jump into code:** Ready to build your own scripts? Here's an example querying for Pacific Cod in the Gulf of Alaska for 2021:
+
+```
+import afscgap  # install with pip install afscgap
+query = afscgap.Query()
+query.filter_year(eq=2021)
+query.filter_srvy(eq='GOA')
+query.filter_scientific_name(eq='Pasiphaea pacifica')
+results = query.execute()
+```
+
+Continue your exploration in the [developer docs]().
 
 <br>
 <br>
