+++ afscgapviz/static/js/intro.js
@@ -24,7 +24,7 @@ const STEP_TEXTS = [
     "The temperatures in some areas of the Gulf of Alaska appear to have been higher at the time of haul compared to the Aleutian Islands.",
     "Was this also true right before The Blob? Let's compare the Gulf of Alaska in 2013 to the Aleutian Islands in 2012.",
     "Looks like it! There are areas with higher temperatures in the Gulf of Alaska compared to the Aleutian Islands beforehand as well.",
-    "That in mind, to better see what happened with The Blob, let's overlay the data. On the Gulf of Alaska display, let's put Pacific cod 2013 as Scatter 1 and Pacific cod 2015 as Scatter 2. Then, let's do 2014 for Scatter 1 and 2016 for Scatter 2 in the Aleutian Islands. Note that the year selector will appear after selecting a species.",
+    "That in mind, to better see what happened with The Blob, let's overlay the data. On the Gulf of Alaska display, let's put Pacific cod 2013 as Scatter 1 and Pacific cod 2015 as Scatter 2. Then, let's do 2014 for Scatter 1 and 2016 for Scatter 2 in the Aleutian Islands. Note that the year selector will appear after typing in a species where it currently says None.",
     "With these temperature deltas visualized, how much did temperature change in each region? How did the catch change in areas of warming?",
     "For Pacific cod, <a target='_blank' href='https://cdnsciencepub.com/doi/full/10.1139/cjfas-2019-0238'>research suggests that warming matters quite a lot in a region but it's a complex phenomenon impacting spawning habitat</a>. The Gulf of Alaska with widespread warming may have been under more pressure than the Aleutian Islands.",
     "This is a lot of info on Pacific cod. How about walleye pollock? Did that species see something similar happen? Go ahead and explore that species below.",
