### 2023-04-28 fe929ca
Force refresh on focus out
