### 2023-04-28 34ddb5c
Update install browser script for Firefox.
