+++ install_browser.sh
@@ -1,1 +1,4 @@
-sudo apt install -y chromium-browser
+sudo apt-key adv --keyserver keyserver.ubuntu.com --recv-keys A6DCF7707EBC211F
+sudo apt-add-repository "deb http://ppa.launchpad.net/ubuntu-mozilla-security/ppa/ubuntu bionic main"
+sudo apt-get update
+sudo apt-get install firefox
