### 2023-05-01 e1e9049
Revert "Update install browser script for Chrome."
