+++ install_browser.sh
@@ -1,2 +1,1 @@
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo dpkg -i google-chrome-stable_current_amd64.deb
+sudo apt install -y chromium-browser
