### 2023-05-01 402ceb8
Make download filename more descriptive.
Closes #60
