### 2023-05-04 #93 @ocefpaf
PS: Don't forget to [close your review](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template) with the reviewer template. See https://github.com/pyOpenSci/software-submission/issues/93#issuecomment-1507160030.
