### 2023-05-04 #93 @sampottinger
Hello @ayushanand18! Thanks very much for your review.

> However, my only concern is of including Installation Instructions for the dev version of the package through GitHub.

I’m sorry are you recommending that we include additional installation details for installing direct from GitHub (eg: `pip install git+https://github.com/SchmidtDSE/afscgap.git@main`)? Just wanted to double check.
