### 2023-05-04 #93 @sampottinger
Also @ocefpaf / @ayushanand18  - we were hoping to try at JOSS (we know they don’t like API wrappers but we are hoping that zero catch inference, ORDS compilation, visualization, etc take us over the line). The `paper.md` is in the `inst` folder. LMK if there is something else we should do to facilitate. Just noticed that the section on JOSS was left unchecked.
