### 2023-05-04 #93 @sampottinger
Sorry @ayushanand18 - Added this PR to address your final comment: https://github.com/SchmidtDSE/afscgap/pull/72. LMK if this addresses your concern. Thanks again very much for your time! We very much appreciate all of your help.
