### 2023-05-05 #93 @ayushanand18
> I’m sorry

No need to be sorry on that, I should have been clearer in my comments :)

> are you recommending that we include additional installation details for installing direct from GitHub

Yes, I meant that. The PR looks good, and I don't have anything to add.
