### 2023-05-09 #93 @7yl4r
Sorry for the delay on my part. I have filled the review template in [this gist](https://gist.github.com/7yl4r/1e2e124e696a06b4f52eb61fbdeae27c) but am still looking through things. All looks great. There are some specific details from pyOpenSci to look at. My personal preference would be to simplify the documentation text, but I do not yet have a recommendation on how to do that.
