### 2023-05-15 SchmidtDSE/afscgap#75 @7yl4r
Jupyter notebook cartopy installation 
Trying to run the Jupyter notebook from Ubuntu 22.04.2 LTS and I am seeing an error w/ Cartopy usage:

![image](https://github.com/SchmidtDSE/afscgap/assets/1051390/d35ac746-6db6-456c-a572-4f7ad3ea277d)

Steps taken to get here:
```
$ git checkout git@github.com:SchmidtDSE/afscgap.git
$ cd afscgap/notebooks
$ pip install -r requirements.txt
$ jupyter notebook
```

Looks like pip didn't fully set up what was needed?
