### 2023-05-15 SchmidtDSE/afscgap#75 @7yl4r
`sudo apt -y install libgeos-dev` (as suggested in [cartopy install docs](https://scitools.org.uk/cartopy/docs/latest/installing.html#ubuntu-debian) does not fix.
