### 2023-05-16 d142bd3
Remove think-aloud
Not all participants used it.
