+++ inst/paper.bib
@@ -193,14 +193,6 @@
   year={2023}
 }
 
-@misc{thinkaloud,
-  title={Using the "Thinking-aloud" method in cognitive interface design},
-  author={Lewis, Clayton},
-  year={1982},
-  publisher={IBM TJ Watson Research Center},
-  url={https://dominoweb.draco.res.ibm.com/2513e349e05372cc852574ec0051eea4.html}
-}
-
 @misc{toolz,
   title={Pytoolz API Documentation},
   url={https://toolz.readthedocs.io/en/latest/},
