+++ inst/paper.md
@@ -82,7 +82,7 @@ Of course, building competency in a sophisticated interface like this presents u
 
 While this interface uses game / information design techniques to offer an accessible on-ramp to quickly learn a sophisticated interface, it also serves as a starting point for continued analysis by generating either CSV or Python code to take work into other tools.
 
-Note that, in addition to use in a graduate classroom lecture setting, five experts with related scientific background have provided tool-specific feedback in a quality / usability assurance capacity. Four of the five provided feedback by video conference while one offered commentary by email.
+Note that, in addition to use in a graduate classroom lecture setting, five individuals with related scientific background have provided tool-specific feedback in a quality / usability assurance capacity. Four of the five provided feedback by video conference while one offered commentary by email.
 
 ## Limitations
 Notable current limitations:
