### 2023-05-16 8766d5e
Further shortened paper.
