+++ inst/paper.md
@@ -46,18 +46,18 @@ Working with these data requires knowledge of tools ouside the Python "standard 
 Access to the API alone cannot support some investigations as the API provides "presence-only" data [@inport]. Many types of analysis require information not just about where a species was present but also where it was not. For example, consider geohash-aggregated species catch per unit effort: while the presence-only dataset may provide a total weight or count for a species, the total area swept must include hauls in which the speices was not found [@geohash; @notebook]. The `afscgap` Python package can, with memory efficiency, algorithmically infer those "zero catch" records.
 
 ## General public accessibility
-Though the `afscgap` Python package makes GAP catch data more accessible to developers, the size and complexity of this dataset requires non-trivial engineering for comparative analysis between species, years, and/or geographic areas [@notebook]. Without deep developer experience, it can be difficult to get started. To address a broader audience, this project also offers a no-code visualization tool sitting on top of `afscgap` to begin investigations with CSV and Python code export as a bridge to further analysis.
+Though the `afscgap` Python package makes GAP catch data more accessible to developers, the size and complexity of this dataset complicates comparative analysis between species, years, and/or geographic areas [@notebook]. Without deep developer experience, it can be difficult to get started. To address a broader audience, this project also offers a no-code visualization tool sitting on top of `afscgap` to begin investigations with CSV and Python code export as a bridge to further analysis.
 
 # Functions
 This project aims to improve accessibility of GAP catch data and offer inclusive approachable tools to kickstart analysis.
 
 ## Lazy querying facade
-The `afscgap` library manages significant data structure complexity to offer a simple familiar interface to Python developers. First, lazy "generator iterables" increase accessibility by encapsulating logic for memory-efficient pagination and "data munging" behind Python-standard iterators [@lazy]. Furthermore, to support zero catch data, decorators adapt diverse structures to common interfaces, offering polymorphism [@decorators]. Finally, offering a single object entry-point into the library, a "facade" approach frees users from needing deep understanding of the library's types, a goal furthered by compilation of "standard" Python types to Oracle REST Data Service queries [@facade].
+The `afscgap` library manages significant complexity to offer a simple familiar interface to Python developers. First, lazy "generator iterables" increase accessibility by encapsulating logic for memory-efficient pagination and "data munging" behind Python-standard iterators [@lazy]. Furthermore, to support zero catch data, decorators adapt diverse structures to common interfaces, offering polymorphism [@decorators]. Finally, offering a single object entry-point into the library, a "facade" approach frees users from needing deep understanding of the library's types, a goal furthered by compilation of "standard" Python types to Oracle REST Data Service queries [@facade].
 
 ![Diagram of simplified afscgap operation [@diagrams].\label{fig:library}](library.png)
 
 ## Zero catch inference
-"Negative" or "zero catch" inference enables scientists to conduct a broader range of analysis. To achieve this, the package uses the following algorithm:
+"Negative" or "zero catch" inference enables a broader range of analysis. To achieve this, the package uses the following algorithm:
 
  - Paginate while records remain available from the API service.
    - Record species and hauls observed from API-returned results.
@@ -69,33 +69,25 @@ The `afscgap` library manages significant data structure complexity to offer a s
 Note that, this library offers Python-emaultion of ORDS-compiled fitlers for inferred records.
 
 ## Visualization
[…]
