### 2023-05-16 SchmidtDSE/afscgap#75 @sampottinger
Thanks @7yl4r! Unfortunately this is working for me on PopOS. That said, [our binder](https://mybinder.org/v2/gh/SchmidtDSE/afscgap/main?urlpath=/tree/index.ipynb) is actually a Docker container which uses `ubuntu:jammy-20230301` as its base. I just re-ran that in Docker and Cartopy appears to work there. You can see the docker file at https://github.com/SchmidtDSE/afscgap/blob/main/Dockerfile. Can you see if that Docker container works for you? Here is how I ran it...

```
$ git checkout git@github.com:SchmidtDSE/afscgap.git
$ cd afscgap
$ docker build --no-cache -t afscgap-notebooks .
$ docker run -it --rm -p 8888:8888 afscgap-notebooks jupyter notebook --ip=0.0.0.0 --port=8888
```

So, it does appear to work for me under a "fresh" Ubuntu VM. Here's the equivalent of what the docker build does:

```
$ sudo apt-get update
$ sudo apt-get install -y python3-pip
$ sudo apt-get install -y python-is-python3
$ pip install --no-cache --upgrade pip
$ sudo apt-get -y install libgeos-dev
$ sudo pip install -r /requirements.txt
```

All that said, please let me know if there is anything else I can do to further assist. However, if the issue appears to be Cartopy related, I encourage you to reach out to that project directly in their [issues tracker](https://github.com/SciTools/cartopy/issues). Thanks!
