### 2023-05-16 #93 @sampottinger
Hello @7yl4r! Thank you for https://github.com/SchmidtDSE/afscgap/issues/75. I have responded there but I was unable to reproduce your behavior in a fresh Ubuntu VM. Just checking in... is there anything else I can help with at this time? Thank you!
