### 2023-05-17 SchmidtDSE/afscgap#75 @ocefpaf
Wheels for cartopy have been a thorny issue for a while and building your own is not easy due to the geos+shapely dependencies conflicts. However, lots of progress have been made lately and looks like the next version of cartopy will use only shapely 2.0 and drop the direct dependency on geos. With that said, it is not a problem for `afscgap` to solve. Maybe just add some notes on the docs recommending a pre-built cartopy from a 3rd party packager until this situation improves.
