### 2023-05-17 6a62cc0
Expand notebooks usage docs
Closes #75
