### 2023-05-19 #93 @7yl4r
- [ ] ~~Any additional setup required to use the package (authentication tokens, etc.)~~
- [ ] ~~If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.~~

- [ ] ~~**Performance:** Any performance claims of the software been confirmed.~~

---

Just a few general comments on my experince digging into this package:
* The README is long; I have a personal preference for a very terse README that links out to the heavy documentation.
* Badges should technically be above the package description according to review template, but I don't see this as important.
* Overall I found the documentation to be extremely detailed - perhaps overwhelmingly so. Ensuring the quickstart examples and notebooks are front and center is very important. The readme already does a good job of this, but perhaps links out to more usage examples would be beneficial.
