+++ wiki/inference.md
@@ -1,0 +1,112 @@
+# Absence vs presence data
+The API itself provides access to presence only data. This means that records are only given for when a species was found. This can cause issues if trying to aggregate data like, for example, to determine the weight of the species in a region in terms of catch weight per hectare. The AFSC GAP API on its own would not necessarily provide the total nubmer of hecatres surveyed in that region because hauls without the species present would be excluded. That in mind, this library provides a method for inferring absence data.
+
+<br>
+<br>
+
+## Example of absence data in aggregation
+Here is a practical memory efficient example using [geolib](https://pypi.org/project/geolib/) and [toolz](https://github.com/pytoolz/toolz) to aggregate catch data by 5 character geohash.
+
+```
+import afscgap
+import geolib.geohash
+import toolz.itertoolz
+
+import afscgap
+
+query = afscgap.Query()
+query.filter_year(eq=2021)
+query.filter_srvy(eq='GOA')
+query.filter_scientific_name(eq='Gadus macrocephalus')
+query.set_presence_only(False)
+results = query.execute()
+
+def simplify_record(full_record):
+    latitude = full_record.get_latitude(units='dd')
+    longitude = full_record.get_longitude(units='dd')
+    geohash = geolib.geohash.encode(latitude, longitude, 5)
+    
+    return {
+        'geohash': geohash,
+        'area': full_record.get_area_swept(units='ha'),
+        'weight': full_record.get_weight(units='kg')
+    }
+
+def combine_record(a, b):
+    assert a['geohash'] == b['geohash']
+    return {
+        'geohash': a['geohash'],
+        'area': a['area'] + b['area'],
+        'weight': a['weight'] + b['weight']
+    }
+
+simplified_records = map(simplify_record, results)
+totals_by_geohash = toolz.reduceby(
+    'geohash',
+    combine_record,
+    simplified_records
+)
+weight_by_area_tuples = map(
+    lambda x: (x['geohash'], x['weight'] / x['area']),
+    totals_by_geohash.values()
+)
+weight_by_area_by_geohash = dict(weight_by_area_tuples)
+```
+
+For more details see the [Python functional programming guide](https://docs.python.org/3/howto/functional.html). All that said, for some queries, the use of Pandas may lead to very heavy memory usage.
+
+<br>
+
+## Absence inference algorithm
+Though it is not possible to resolve this issue using the AFSC GAP API service alone, this library can infer those missing records using a separate static flat file provided by NOAA and the following algorithm:
+
+
+ - Record the set of species observed from API service returned results.
+ - Record the set of hauls observed from API service returned results.
+ - Return records normally while records remain available from the API service.
+ - Upon exhaustion of the API service results, [download the ~10M hauls flat file](https://pyafscgap.org/community/hauls.csv) from this library's community.
+ - For each species observed in the API returned results, check if that species had a record for each haul reported in the flat file.
+ - For any hauls without the species record, yield an 0 catch record from the iterator for that query.
+
+This procedure is disabled by default. However, it can be enabled through the `presence_only` keyword in `query` like so: `asfcgap.query(presence_only=False)`.
+
+<br>
+
+## Memory efficiency of absence inference
[…]
