### 2023-05-19 ab6470f
Switch to build docs script.
