### 2023-05-19 6704a1b
Move install step for docs into build_docs.
