### 2023-05-19 #93 @sampottinger
Thank you all! @7yl4r - I want you to know that I took your suggestions to heart and I am working on a PR to address them at https://github.com/SchmidtDSE/afscgap/pull/81. I should have that resolved later today.
