+++ docs/usage.md
@@ -1,5 +1,7 @@
-# Getting started
+# Usage
 [After you have installed the package](https://pyafscgap.org/devdocs) through `pip install afscgap` or similar, this documentation leverages short examples in a tutorial-like setting to show how to use the `afscgap` package.
+
+Note that [full formalized API documentation is available](https://pyafscgap.org/devdocs/afscgap.html) as generated by pdoc in CI / CD.
 
 <br>
 <br>
