+++ docs/usage.md
@@ -1,5 +1,5 @@
 # Usage
-[After you have installed the package](https://pyafscgap.org/devdocs) through `pip install afscgap` or similar, this documentation leverages short examples in a tutorial-like setting to show how to use the `afscgap` package.
+[After you have installed the package](https://pyafscgap.org/docs/##installation) through `pip install afscgap` or similar, this documentation leverages short examples in a tutorial-like setting to show how to use the `afscgap` package.
 
 Note that [full formalized API documentation is available](https://pyafscgap.org/devdocs/afscgap.html) as generated by pdoc in CI / CD.
 
@@ -57,12 +57,12 @@ for record in results:
 print(count_by_temperature_c)
 ```
 
-See [data structure section](#data-structure). Using an iterator will have the library negotiate pagination behind the scenes so this operation will cause multiple HTTP requests while the iterator runs.
+See [data structure section](https://pyafscgap.org/docs/#model/). Using an iterator will have the library negotiate pagination behind the scenes so this operation will cause multiple HTTP requests while the iterator runs.
 
 <br>
 
 ## Enable absence data
-One of the major limitations of the official API is that it only provides presence data. However, this library can optionally infer absence or "zero catch" records using a separate static file produced by NOAA AFSC GAP. The [algorithm and details for absence inference](#absence-vs-presence-data) is further discussed below.
+One of the major limitations of the official API is that it only provides presence data. However, this library can optionally infer absence or "zero catch" records using a separate static file produced by NOAA AFSC GAP. The [algorithm and details for absence inference](https://pyafscgap.org/docs/#inference/) is further discussed below.
 
 Absence data / "zero catch" records inference can be turned on by passing `False` to `set_presence_only` in `Query`. To demonstrate, this example finds total area swept and total weight for Gadus macrocephalus from the Aleutian Islands in 2021:
 
@@ -90,7 +90,7 @@ message = template % (weight_per_area, total_weight, total_area)
 print(message)
 ```
 
-For more [details on the zero catch record feature](#absence-vs-presence-data), please see below.
+For more [details on the zero catch record feature](https://pyafscgap.org/docs/#inference/), please see below.
 
 <br>
 
