+++ docs/limitations.md
@@ -5,7 +5,7 @@ There are a few caveats for working with these data that are important for resea
 <br>
 
 ## Incomplete or invalid records
-Metadata fields such as `year` are always required to make a `Record` whereas others such as catch weight `cpue_kgkm2` are not present on all records returned by the API and are optional. See the [data structure section](#data-structure) for additional details. For fields with optional values:
+Metadata fields such as `year` are always required to make a `Record` whereas others such as catch weight `cpue_kgkm2` are not present on all records returned by the API and are optional. See the [data structure section](https://pyafscgap.org/docs/#model/) for additional details. For fields with optional values:
 
  - A maybe getter (like `get_cpue_weight_maybe`) is provided which will return None without error if the value is not provided or could not be parsed.
  - A regular getter (like `get_cpue_weight`) is provided which asserts the value is not None before it is returned.
