+++ mkdocs.yml
@@ -8,3 +8,4 @@ nav:
   - Data Model: model.md
   - Inference: inference.md
   - Limitations: limitations.md
+  - Building: building.md
