### 2023-05-19 126a2e4
Simplify build script.
