+++ build_docs.sh
@@ -13,6 +13,5 @@ cd ..
 
 echo "[4 / 4] Build mkdocs..."
 cp README.md docs/index.md
-mkdocs build
-mv site website/docs
+mkdocs build 
 [[ ! -f website/docs/index.html ]] && exit 1
