+++ mkdocs.yml
@@ -1,6 +1,7 @@
 site_name: Python afscgap package
 site_url: https://pyafscgap.org/docs
 theme: windmill
+site_dir: website/docs
 nav:
   - Home: index.md
   - Usage: usage.md
