### 2023-05-19 3c61d26
Remove unneeded build dir
