### 2023-05-19 #93 @sampottinger
We are all set. Please let me know how you want to proceed @ocefpaf. Thanks @7yl4r for your feedback! The README is refactored and the docs microsite reorganized at https://pyafscgap.org/docs/#
