+++ README.md
@@ -156,7 +156,7 @@ Thank you to all of these projects for their contribution.
 ## Version history
 Annotated version history:
 
- - `1.0.0`: Release with pyopensci.
+ - `1.0.0`: (Pending) Release with pyopensci.
  - `0.0.9`: Fix with issue for certain import modalities and the `http` module.
  - `0.0.8`: New query syntax (builder / chaining) and units conversions.
  - `0.0.7`: Visual analytics tools.
