+++ inst/paper.bib
@@ -38,7 +38,7 @@
   url={https://github.com/jgraph/drawio},
   journal={GitHub},
   publisher={GitHub, Inc}, 
-  author={JGraph Ltd and draw.io AG},
+  author={JGraph and draw.io},
   year={2023}
 }
 
