+++ inst/paper.md
@@ -85,10 +85,10 @@ This library offers Python-emulation of ORDS filters for inferred records.
 ## Visualization
 Despite these developer-focused tools, this dataset's complexity requires technical sophistication to navigate. To further increase accessibility, this project offers a visualization tool for starting temporal, spatial, and species comparisons with coordinated highlighting, separated color channels, summary statistics, and side-by-side display [@few]. To support learning this UI, an introduction sequence tutorializes a "real" analysis via Hayashida level design [@hayashida; @brown]:
 
- - **Introduction**: The player sees information about Pacific cod with pre-filled elements used to achieve that analysis gradually fading in.
- - **Development**: Using the mechanics introduced moments prior, the tool invites the player to change the analysis to compare different regions with temperature data.
- - **Twist**: Overlays on the same display are enabled, allowing the player to leverage mechanics they just exercised in a now more complex interface.
- - **Conclusion**: End with giving the player an opportunity to demonstrate all of the skills acquired in a new problem.
+ - **Introduction**: The tool shows information about Pacific cod with pre-filled elements used to achieve that analysis gradually fading in.
+ - **Development**: Using the mechanics introduced moments prior, the tool invites the user to change the analysis to compare different regions with temperature data.
+ - **Twist**: Overlays on the same display are enabled, allowing the user to leverage mechanics they just exercised in a now more complex interface.
+ - **Conclusion**: End with giving the user an opportunity to demonstrate all of the skills acquired in a new problem.
 
 Note that this visualization also serves as a starting point for continued analysis by generating either CSV or Python code to take work into other tools.
 
