### 2023-05-21 1bc14f8
Minor modifications ahead of JOSS.
