+++ afscgapviz/static/js/intro.js
@@ -19,10 +19,9 @@ const STEP_TEXTS = [
     "Specifically, let's look at temperature readings taken at the bottom by selecting \"Bottom temperatures\" for the first display below. We will add it to the other display in a moment.",
     "Now, let's do the same on the other display. Go ahead and select bottom temperatures again in the other dropdown. How did they change between years?",
     "It looks like there may have been some warming in the Gulf of Alaska. I wonder if this was region-wide?",
-    "What about the Aleutian Islands? Did the same temperatures appear there? Try putting Gulf of Alaska data in one display and Aleutian Islands data in the other. Again, let's stick with Pacific cod for now.",
-    "Specifically, go ahead and select 2015 in the Gulf of Alaska and 2014 in the Aleutian Islands (Pacific cod in both). Note that a survey does not happen in every region every year.",
+    "What about the Aleutian Islands? Did the same temperatures appear there? Try putting Gulf of Alaska data in one display and Aleutian Islands data in the other. Specifically, select 2015 in the Gulf of Alaska and 2014 in the Aleutian Islands (Pacific cod in both). Note that a survey does not happen in every region every year.",
     "The temperatures in some areas of the Gulf of Alaska appear to have been higher at the time of haul compared to the Aleutian Islands.",
-    "Was this also true right before The Blob? Let's compare the Gulf of Alaska in 2013 to the Aleutian Islands in 2012.",
+    "Was this also true right before The Blob? Compare the Gulf of Alaska in 2013 to the Aleutian Islands in 2012.",
     "Looks like it! There are areas with higher temperatures in the Gulf of Alaska compared to the Aleutian Islands beforehand as well.",
     "That in mind, to better see what happened with The Blob, let's overlay the data. On the Gulf of Alaska display, let's put Pacific cod 2013 as Scatter 1 and Pacific cod 2015 as Scatter 2. Then, let's do 2014 for Scatter 1 and 2016 for Scatter 2 in the Aleutian Islands. Note that the year selector will appear after typing in a species where it currently says None.",
     "With these temperature deltas visualized, how much did temperature change in each region? How did the catch change in areas of warming?",
@@ -70,7 +69,6 @@ class Intro {
             [".temperature-select-2"],
             [],
             [".area-select", ".species-type-select"],
-            [],
             [],
             [],
             [],
