+++ inst/paper.md
@@ -108,10 +108,9 @@ Thanks to:
 
  - Runtime dependencies: ColorBrewer, D3, Flask, Geolib, Requests, Toolz, and Papa Parse [@colorbrewer; @d3; @flask; @geolib; @requests; @toolz; @papa].
  - Library advice: Carl Boettiger, Fernando Perez, and PyOpenSci reviewers.
- - Visualization advice: Magali de Bruyn, Brookie Guzder-Williams, Angela Hayes, David Joy, Maya Weltman-Fahs, and Lewis Barnett.
+ - Visualization advice: Lewis Barnett, Magali de Bruyn, Brookie Guzder-Williams, Angela Hayes, David Joy, Emily Markowitz, and Maya Weltman-Fahs.
  - Ciera Martinez for general guidance.
  - Draw.io for diagrams [@diagrams].
- - Others declining acknowledgment.
 
 Project of The Eric and Wendy Schmidt Center for Data Science and the Environment at UC Berkeley.
 
