### 2023-05-31 #93 @sampottinger
Thanks for your help @ocefpaf! I got everything done today except the JOSS submission form. Things should be good to go on the pyopensci side though. I'll try to submit to JOSS tomorrow. Thanks!
