+++ build_docs.sh
@@ -8,7 +8,7 @@ pdoc --docformat google ./afscgap -o website/devdocs/
 echo "[3 / 4] Build paper preview..."
 cd inst
 bash preview_paper.sh
-[[ ! -f ../website/static/paper.pdf ]] && exit 1
+[[ ! -f ../website/static/paper_preview.pdf ]] && exit 1
 cd ..
 
 echo "[4 / 4] Build mkdocs..."
