+++ .github/workflows/build.yml
@@ -54,3 +54,14 @@ jobs:
         uses: actions/checkout@v2.3.4
       - name: Build docs
         run: bash build_docs.sh
+  joss:
+    name: JOSS Draft
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Build draft PDF
+        uses: openjournals/openjournals-draft-action@master
+        with:
+          journal: joss
+          paper-path: paper/paper.md
