+++ build_docs.sh
@@ -6,7 +6,7 @@ pdoc --docformat google ./afscgap -o website/devdocs/
 [[ ! -f website/devdocs/afscgap.html ]] && exit 1
 
 echo "[3 / 4] Build paper preview..."
-cd inst
+cd paper
 bash preview_paper.sh
 [[ ! -f ../website/static/paper_preview.pdf ]] && exit 1
 cd ..
