### 2023-06-01 a58b0e1
Leave the JOSS version as an artifact.
