+++ paper/README.md
@@ -1,6 +1,6 @@
 Paper
 ================================================================================
-Resources for the JOSS paper submission for the Pyafscgap.org project. See paper preview at https://pyafscgap.org/static/paper_preview.pdf or the JOSS formatted version at https://pyafscgap.org/static/paper_joss.pdf. Note that the paper preview uses the following open source tools:
+Resources for the JOSS paper submission for the Pyafscgap.org project. See paper preview at https://pyafscgap.org/static/paper_preview.pdf or the JOSS formatted version in the build artifact. Note that the paper preview uses the following open source tools:
 
  - [pandoc/luafilters](https://github.com/pandoc/lua-filters) by pandoc Lua filters contributors under the [MIT License](https://github.com/pandoc/lua-filters/blob/master/LICENSE)
  - [pandoc](https://pandoc.org/) (via call to OS package through shell script) by John MacFarlane, Albert Krewinkel, Jesse Rosenthal, and other contributors under the [GPL License](https://github.com/jgm/pandoc#license).
