### 2023-06-01 51678b3
Move the visualization screenshot forward.
