### 2023-06-01 #93 @lwasser
congratulations @sampottinger and welcome to our ecosystem of packages!! 🚀 i'll make sure your pr's are merged today so you get added to our website!! ✨ please keep us posted on the joss submission as well as we will update our tags. we also appreciate referencing this issue when you submit so they two reviews are connected (and they know not to actually re-review your package!!). @ocefpaf thank you for leading a really efficient review! and  @7yl4r @ayushanand18 thank you for reviewing for us!!! we will add you to our website as contributors as well.

last request: if you all have the bandwidth to [take our post review survey ](https://forms.gle/4eV72RWW3JnEkPoh8) id GREATLY appreciate it. we use this feedback on the review to track package improvements and also to just get feedback so we can improve over time. thank you in advance for doing this.
