### 2023-06-01 50b57ac
Added missing person on acknowledgements.
Since we just had a doc update release, I won't do another but saw we were missing David.
