### 2023-06-02 923b0f6
Remove unneeded text in viz section.
