### 2023-06-02 #93 @ocefpaf
@sampottinger let me know if you need any help with the zenodo DOI and badge. Those are the last boxes we can tick in https://github.com/pyOpenSci/software-submission/issues/93#issuecomment-1568912956 that do not require waiting on JOSS :-)
