### 2023-06-02 7d8a55a
Minor paper touchups.
