+++ paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-The Groundfish Assessment Program within NOAA's Alaska Fisheries Science Center produces longitudinal catch data for the region [@afscgap]. Supporting ocean health research and fisheries management, these "hauls" report where marine species are found during bottom trawl surveys and in what quantities [@example]. Increasing data usability for communities of diverse programming experience, Pyafscgap.org offers not just friendlier REST API operation for this economically and scientifically important dataset but query language compliation, memory-efficient algorithms for "zero-catch" inference, and interactive visual analytics. Altogether, this research toolset supports investigatory tasks not easily executable using the API service alone and broadens access through game and information design.
+The Groundfish Assessment Program within NOAA AFSC produces longitudinal catch data [@afscgap]. Supporting ocean health research and fisheries management, these "hauls" report where bottom trawl surveys find marine species and in what quantities [@example]. Increasing these data usability for communities of diverse programming experience, Pyafscgap.org offers query language compliation, memory-efficient algorithms for "zero-catch" inference, and interactive visual analytics for multiple locations' economically and scientifically important datasets. Altogether, this research toolset supports investigatory tasks across survey programs and broadens access through game and information design.
 
 # Statement of need
 Pyafscgap.org reduces barriers for use of NOAA AFSC RACE GAP^[Groundfish Assessment Program in the Resource Assessment and Conservation Engineering Division of the National Oceanic and Atmospheric Administration's Alaska Fisheries Science Center] data, offering:
@@ -43,17 +43,17 @@ Pyafscgap.org reduces barriers for use of NOAA AFSC RACE GAP^[Groundfish Assessm
  - Memory-efficient algorithms for analysis requiring zero catch inference.
  - Tools for those with less developer experience.
 
-Altogether, these open source tools extend the dataset's reach and approachability.
+Altogether, these open source tools extend the reach and approachability of the multiple survey programs operating under the division. Example utilization includes longitudinal analysis of catch per unit effort (CPUE) in context of fisheries management and environmental changes [@notebook].
 
 ## Developer usability
-Working with these data requires knowledge of tools ouside the Python "standard toolset" like the closed-source Oracle REST Data Services (ORDS) [@ords]. The `afscgap` package offers easier access to the official REST service with automated pagination, ORDS compilation, and documented types. Together, these tools enable Python developers to use familiar patterns to interact with these data: type checking, standard documentation, and compatability with common Python data-related libraries.
+Working with these data requires knowledge of tools ouside the Python "standard toolset" like closed-source ORDS query language [@ords]. While the `afscgap` package offers easier access to the official REST service, it also crucially offers ORDS compilation, documented types, and lazy access to these sometimes large datasets. Together, these tools enable Python developers to use familiar patterns to efficiently interact with these data: type checking, standard documentation, lazy access, and compatability with common Python data-related libraries.
 
 ## Record inference
-The API struggles supporting some investigations as it provides "presence-only" data [@inport]. For example, the API may readily yield total mass of Pacific cod but not its geohash-aggregated catch per unit effort [@geohash].
+The surveys on their own within the API struggle supporting some investigations as they provide "presence-only" data [@inport]. For example, the API may readily yield total mass of Pacific cod but not its geohash-aggregated CPUE [@geohash].
 
 $$CPUE_{species} = \frac{m_{species}}{A_{swept}}$$
 
-Knowing CPUE (kg/ha) must also include "absence data" or hauls in which the speices was not recorded, this package can efficiently infer those hauls not easily returned from the API service [@notebook].
[…]
