### 2023-06-05 54cda0d
Minor simplification.
