### 2023-06-17 3e22f91
Just use open journals draft action.
