+++ build_docs.sh
@@ -1,17 +1,11 @@
-echo "[1 / 4] Setup environment..."
+echo "[1 / 3] Setup environment..."
 pip install -e .[dev]
 
-echo "[2 / 4] Build API documentation..."
+echo "[2 / 3] Build API documentation..."
 pdoc --docformat google ./afscgap -o website/devdocs/
 [[ ! -f website/devdocs/afscgap.html ]] && exit 1
 
-echo "[3 / 4] Build paper preview..."
-cd paper
-bash preview_paper.sh
-[[ ! -f ../website/static/paper_preview.pdf ]] && exit 1
-cd ..
-
-echo "[4 / 4] Build mkdocs..."
+echo "[3 / 3] Build mkdocs..."
 cp README.md docs/index.md
 mkdocs build
 [[ ! -f website/docs/index.html ]] && exit 1
