+++ paper/paper.md
@@ -25,16 +25,16 @@ author:
       institute:
         - ucberkeley
 affiliations:
- - name: University of California, Berkeley
+ - name: University of California Berkeley, California, United States of America
    index: 1
 institute:
-  - ucberkeley: University of California, Berkeley
+  - ucberkeley: University of California Berkeley, California, United States of America
 date: 2 June 2023
 bibliography: paper.bib
 ---
 
 # Summary
-NOAA AFSC's Groundfish Assessment Program produces longitudinal catch data [@afscgap]. Supporting ocean health research and fisheries management, these "hauls" report in what quantities, conditions, and locations bottom trawl surveys find different marine species [@example]. Increasing usability for communities of diverse programming experience, Pyafscgap.org offers query language compliation, memory-efficient algorithms for "zero-catch" inference, and interactive visual analytics for these economically and scientifically important GAP datasets. Altogether, this research toolset supports investigatory tasks across survey programs' locations and broadens access through game and information design.
+NOAA AFSC's Groundfish Assessment Program produces longitudinal catch data [@afscgap]. Supporting ocean health research and fisheries management, these "hauls" report in what quantities, conditions, and locations bottom trawl surveys find different marine species [@example]. Increasing usability for communities of diverse programming experience, Pyafscgap.org offers query language compilation, memory-efficient algorithms for "zero-catch" inference, and interactive visual analytics for these economically and scientifically important GAP datasets. Altogether, this research toolset supports investigatory tasks across survey programs' locations and broadens access through game and information design.
 
 # Statement of need
 Pyafscgap.org reduces barriers for use of NOAA AFSC RACE GAP^[Groundfish Assessment Program in the Resource Assessment and Conservation Engineering Division of the National Oceanic and Atmospheric Administration's Alaska Fisheries Science Center] data, offering:
@@ -46,14 +46,14 @@ Pyafscgap.org reduces barriers for use of NOAA AFSC RACE GAP^[Groundfish Assessm
 Altogether, these open source tools extend the reach and approachability of GAP's multiple survey programs, supporting analysis like longitudinal catch per unit effort (CPUE) in context of environmental changes [@notebook].
 
 ## Developer usability
-Working with these data requires knowledge of tools ouside the Python "standard toolset" like closed-source ORDS query language [@ords]. While the `afscgap` package offers easier access to the official REST service, it also crucially offers ORDS compilation, documented types, and lazy access to these large datasets. Together, these tools enable Python developers to efficiently use familiar patterns to interact with these data: type checking, standard documentation, and compatability with common Python data-related libraries.
+Working with these data requires knowledge of tools ouside the Python "standard toolset" like closed-source ORDS query language [@ords]. While the `afscgap` package offers easier access to the official REST service, it also crucially offers ORDS compilation, documented types, and lazy access to these large datasets. Together, these tools enable Python developers to efficiently use familiar patterns to interact with these data: type checking, standard documentation, and compatibility with common Python data-related libraries.
 
 ## Record inference
 Surveys on their own within the API struggle supporting some investigations as they provide "presence-only" data [@inport]. For example, the API may readily yield total mass of Pacific cod but not its geohash-aggregated CPUE [@geohash].
 
 $$CPUE_{species} = \frac{m_{species}}{A_{swept}}$$
 
-Metrics like CPUE need "absence data" (hauls in which the speices was not recorded) and this package can efficiently infer those results [@notebook].
+Metrics like CPUE need "absence data" (hauls in which the species was not recorded) and this package can efficiently infer those results [@notebook].
 
 ## Broad accessibility
[…]
