### 2023-06-28 d134c2d
Minor updates in efficient facade section.
