+++ paper/paper.md
@@ -102,8 +102,8 @@ In addition to use in a graduate classroom setting, five individuals with releva
 As further documented in the repository [@readme], these tools:
 
  - Run single-threaded and synchronous.
- - Aggregates hauls as points in visualization due to data limitation.
- - Ignores hauls if entirelly excluded by NOAA.
+ - Aggregate hauls as points in visualization due to data limitation.
+ - Ignore hauls if entirelly excluded by NOAA.
 
 # Acknowledgements
 Thanks to:
