### 2023-06-28 d384c47
Minor typo fix in paper.
