+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "afscgap"
-version = "1.0.3"
+version = "1.0.4"
 authors = [
   { name="A Samuel Pottinger", email="sam.pottinger@berkeley.edu" },
 ]
