### 2023-06-28 #93 @sampottinger
Hey all! We were accepted into JOSS. https://github.com/openjournals/joss-reviews/issues/5593 / https://joss.theoj.org/papers/10.21105/joss.05593
