### 2023-06-28 #93 @ocefpaf
Congrats @sampottinger. Great work! Also, as someone who work with similar things (making sure the data reaches those who need them) I'm very happy to close this as completed and to see it on JOSS.
